% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_transcription_system.R
\name{add_attribute_to_level}
\alias{add_attribute_to_level}
\title{Add attribute to level}
\usage{
add_attribute_to_level(config, level_name, attribute_name)
}
\description{
Add attribute to level
}
\keyword{internal}
