% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_emuR_re-export.R
\name{add_attributeDefinition}
\alias{add_attributeDefinition}
\title{Define a new attribute for a level. (A copy of the emuR function)}
\usage{
add_attributeDefinition(
  emuDBhandle,
  levelName,
  name,
  type = "STRING",
  rewriteAllAnnots = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link[emuR]{load_emuDB}}}

\item{levelName}{name of level}

\item{name}{name of attributeDefinition}

\item{type}{type of attributeDefinition (currently only "STRING")}

\item{rewriteAllAnnots}{should changes be written to file system (_annot.json files) (intended for expert use only)}

\item{verbose}{if set to \code{TRUE}, more status messages are printed}
}
\description{
Add / List / Rename / Remove database operation functions for attribute
definition to / of / from an existing level definition of an emuDB.
Attribute definitions can be viewed as definitions of
parallel labels for the annotational units (ITEMs) of the emuDB. 
Each level definition is required to have at least one 
default attribute definition that has the same name as the level definition
(automatically created by \code{\link[emuR]{add_levelDefinition}}). For more 
information on the structural elements of an emuDB see \code{vignette(emuDB)}.
Note that as with level definitions, an attribute definition to a level cannot be removed,
if it contains labels in the emuDB.

As the only one of these operations, \code{rename_attributeDefinition} can
also be used to manipulate (i.e. rename) a level definition. It is therefore
not necessary to specify the name of the level that the attribute definition
belongs to. While renaming a level or attribute definition, emuR will
(1) rewrite the levelDefinitions in DBconfig, (2) rewrite the
linkDefinitions in DBconfig, (3) rewrite the perspectives in DBconfig,
(4) rewrite the anagestConfig in DBconfig, and (5) rewrite all _annot.json
files. (5) May take quite a while, depending on the number of bundles in the
database.
}
\seealso{
\link[emuR:AddListRenameRemoveAttributeDefinitions]{emuR::add_attributeDefinition}
}
