% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_transcription_system.R
\name{add_level_to_config}
\alias{add_level_to_config}
\title{Add level to configuration}
\usage{
add_level_to_config(config, level_name, level_type, attribute_name)
}
\description{
Add level to configuration
}
\keyword{internal}
