% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_emuR_re-export.R
\name{add_levelDefinition}
\alias{add_levelDefinition}
\title{Add a level definition. (A copy of the emuR function)}
\usage{
add_levelDefinition(
  emuDBhandle,
  name,
  type,
  rewriteAllAnnots = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link[emuR]{load_emuDB}}}

\item{name}{name of level definition}

\item{type}{type of level definition ("SEGMENT","EVENT","ITEM")}

\item{rewriteAllAnnots}{should changes be written to file system (_annot.json files) (intended for expert use only)}

\item{verbose}{Show progress bars and further information}
}
\description{
Add / List / Remove database operation functions for level definitions. 
A level is a more general term for what is often referred to as a "tier". 
It is more general in the sense that people usually 
expect tiers to contain time information. Levels 
can either contain time information if they are of the 
type "EVENT" or of the type "SEGMENT" but are timeless 
if they are of the type "ITEM". For more information 
on the structural elements of an emuDB see \code{vignette(emuDB)}.
Note that a level cannot be removed, if it contains instances of annotation items
or if it is linked to another level. Further note, renaming a level definition 
can be done using \code{\link[emuR]{rename_attributeDefinition}}.
}
\seealso{
\link[emuR:AddListRemoveLevelDefinitions]{emuR::add_levelDefinition}
}
