% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_emuR_re-export.R
\name{add_linkDefinition}
\alias{add_linkDefinition}
\title{Add a level definition (from emuR)}
\usage{
add_linkDefinition(emuDBhandle, type, superlevelName, sublevelName)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link[emuR]{load_emuDB}}}

\item{type}{type of linkDefinition (either \code{"ONE_TO_MANY"}, \code{"MANY_TO_MANY"} or \code{"ONE_TO_ONE"})}

\item{superlevelName}{name of super-level of linkDefinition}

\item{sublevelName}{name of sub-level of linkDefinition}
}
\description{
Add / List / Remove new link definition to / of / from emuDB. A link definition
specifies the relationship between two levels, the
super-level and the sub-level. The entirety of all link 
definitions of a emuDB specifies the 
hierarchical structure of the database. For more information
on the structural elements of an emuDB see \code{vignette(emuDB)}.
}
\details{
Link type descriptions:
\itemize{
\item{\code{"ONE_TO_MANY"}}{A single ITEM of the super-level can be linked to multiple ITEMs of the sub-level}
\item{\code{"MANY_TO_MANY"}}{Multiple ITEMs of the super-level can be linked to multiple ITEMs of the sub-level}
\item{\code{"ONE_TO_ONE"}}{A single ITEM of the super-level can be linked to a single ITEM of the sub-level}
}

For all link types the rule applies that no links are allowed to cross any other links.
Further, a linkDefinition can not be removed, if there are links present in the emuDB.
}
\seealso{
\link[emuR:AddListRemoveLinkDefinition]{emuR::add_linkDefinition}
}
