% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_emuR_re-export.R
\name{add_perspective}
\alias{add_perspective}
\title{Add a new perspective to an emuDB (from emuR)}
\usage{
add_perspective(emuDBhandle, name)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link[emuR]{load_emuDB}}}

\item{name}{name of perspective}
}
\description{
Add / List / Remove perspective to / of / from emuDB. The EMU-webApp subdivides different ways 
to look at an emuDB into so called perspectives. These perspectives, 
between which you can switch in the web application, contain 
information on what levels are displayed, which ssffTracks are drawn, 
and so on. For more information on the structural elements of an emuDB 
see \code{vignette{emuDB}}.
}
\seealso{
\link[emuR:AddListRemovePerspective]{emuR::add_perspective}
}
