% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_autosync_wrappers.R
\name{add_session_with_sync}
\alias{add_session_with_sync}
\title{Add session with auto-sync}
\usage{
add_session_with_sync(db_handle, name, trigger_sync = TRUE, verbose = FALSE)
}
\arguments{
\item{db_handle}{An emuDBhandle object}

\item{name}{Session name}

\item{trigger_sync}{Logical; trigger auto-sync after adding}

\item{verbose}{Logical; show messages}
}
\description{
Wrapper for emuR::add_session that triggers CMDI sync
}
\keyword{internal}
