% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_cmdi.R
\name{add_speech_corpus_components}
\alias{add_speech_corpus_components}
\title{Add speech corpus components to CMDI}
\usage{
add_speech_corpus_components(
  parent,
  db_name,
  db_uuid,
  corpus_title,
  corpus_description,
  author,
  institution,
  contact_email,
  license,
  availability,
  db_metadata,
  participant_metadata,
  include_placeholders
)
}
\description{
Add speech corpus components to CMDI
}
\keyword{internal}
