% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_emuR_re-export.R
\name{add_ssffTrackDefinition}
\alias{add_ssffTrackDefinition}
\title{Define an SSFF track for the emuDB (from emuR)}
\usage{
add_ssffTrackDefinition(
  emuDBhandle,
  name,
  columnName = NULL,
  fileExtension = NULL,
  onTheFlyFunctionName = NULL,
  onTheFlyParams = NULL,
  onTheFlyOptLogFilePath = NULL,
  verbose = TRUE,
  interactive = TRUE
)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link[emuR]{load_emuDB}}}

\item{name}{name of ssffTrackDefinition}

\item{columnName}{columnName of ssffTrackDefinition.
If the \code{onTheFlyFunctionName} parameter is set and columnName isn't, the
\code{columnName} will default to the first entry in \code{wrasspOutputInfos[[onTheFlyFunctionName]]$tracks}.}

\item{fileExtension}{fileExtension of ssffTrackDefinitions.
If the \code{onTheFlyFunctionName} parameter is set and fileExtension isn't, the
\code{fileExtension} will default to the first entry in \code{wrasspOutputInfos[[onTheFlyFunctionName]]$ext}.}

\item{onTheFlyFunctionName}{name of wrassp function to do on-the-fly calculation. If set to the name of a wrassp 
signal processing function, not only the emuDB schema is extended by the ssffTrackDefintion but also 
the track itself is calculated from the signal file and stored in the emuDB. See \code{names(wrasspOutputInfos)}
for a list of all the signal processing functions provided by the wrassp package.}

\item{onTheFlyParams}{a list of parameters that will be given to the function 
passed in by the onTheFlyFunctionName parameter. This list can easily be 
generated using the \code{\link{formals}} function on the according signal processing function 
provided by the wrassp package and then setting the
parameter one wishes to change.}

\item{onTheFlyOptLogFilePath}{path to optional log file for on-the-fly function}

\item{verbose}{Show progress bars and further information}

\item{interactive}{ask user for confirmation}
}
\description{
Add / List / Remove ssffTrackDefinition to / from / of emuDB. 
An ssffTrack (often simply referred to as a track) references 
data that is stored in the Simple Signal File Format (SSFF) 
in the according bundle folders. The two most common types of data are:
\itemize{
\item{complementary data that was acquired during the recording 
such as data acquired during electromagnetic 
articulographic (EMA) or electropalatography (EPG) recordings;}
\item{derived data, i.e. data that was calculated from the original audio signal 
such as formant values and their bandwidths or the short-term Root Mean Square amplitude of the signal.}
}
For more information on the structural elements of an emuDB see \code{vignette(emuDB)}.
}
\seealso{
\link[emuR:AddListRemoveSsffTrackDefinition]{emuR::add_ssffTrackDefinition}
}
