% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_trackUtilities.R
\name{age_ranges}
\alias{age_ranges}
\title{Cut a vector of ages into age ranges}
\usage{
age_ranges(ages, new.at, suffix = " yrs", start.at.zero = TRUE)
}
\arguments{
\item{ages}{The vector of ages.}

\item{new.at}{A vector of ages at which a new age range should be started,
\emph{or} a single integer indicating the interval where a new interval is
started. For instance, \code{new.at=10} indicates that a range will be
constructed for "20-29 yrs", "30-39 yrs", and so on.}

\item{suffix}{The string suffix that is appended to the age range (defaults
to " yrs"). The user has to add a space to the suffix if required for the
application.}

\item{start.at.zero}{Should the first interval start at zero (the default).}
}
\value{
A factor of age ranges.
}
\description{
The function makes it easier to construct well presented range factors from a
vector of ages, by stating explicitly the first ages that should be included
in a new factor level (as a vector). Alternatively, the used may specify that
a new level should start every \code{new.at} (integer) years.
}
\details{
The function does not perform any rounding up or down, truncating or
any other transformation on the vector of ages before cutting it into
discrete age ranges. So, if the user wants a person of age 10 years and 8
months to be considered part of a "11-13 yrs" age range rather than a
preceding "9-10 yrs" range, then the user should round the ages to whole
integers (e.g. using \code{round(ages,0)} before applying this function.
}
\examples{
ages <- c(1,1.99999,2,4,10,11)
# Unevenly spaced age ranges
age_ranges(ages, new.at=c(0,2,4,6,8,seq(10,130,10)))
# Evenly spaced age ranges
age_ranges(ages, new.at=5)

}
