% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_annotate.R
\name{annotate_INTSINT_MOMEL}
\alias{annotate_INTSINT_MOMEL}
\title{Automatic annotation of intonation using MOMEL and INTSINT labels}
\usage{
annotate_INTSINT_MOMEL(
  emuDBhandle,
  seglist,
  windowSize = 30,
  minF = 60,
  maxF = 750,
  pitchSpan = 1.5,
  maximumError = 1.04,
  reducWindowSize = 20,
  minimalDistance = 20,
  minimalFrequencyRatio = 0.05,
  intsint.level = "Intsint",
  attribute.definitions = c("Momel", "Intsint", "IntsintMomel"),
  absolute.tone.labels = c("T", "M", "B"),
  relative.tone.labels = c("H", "U", "S", "D", "L"),
  force = FALSE,
  verbose = FALSE,
  praat_path = NULL
)
}
\arguments{
\item{emuDBhandle}{The Emu database handle.}

\item{seglist}{A segment list resulting from a \link{query} function call,
containing utterances that should be transcribed.}

\item{windowSize}{The window size (in ms) used when extracting f0. One hald
window size of duration will also be added to each end of all extracted
utterance segments to allow f0 target points to be discovered right at the
beginning and end of utterances.}

\item{minF}{The low end of the fundamental frequency range to consider when
searching for a pitch contour.}

\item{maxF}{The high end of the fundamental frequency range to consider when
searching for a pitch contour.}

\item{pitchSpan}{The maximum pitch span to consider. Defaults to 1.5 octave,
but could be expanded to 2.5 for expressive speech.}

\item{maximumError}{The maximum error allowed by MOMEL. See the momel source
code for more information.}

\item{reducWindowSize}{The size of the reduced analysis window used by MOMEL.
See the momel source code for more information.}

\item{minimalDistance}{The minimal distance between INTSINT labels. See the
momel source code for more information.}

\item{minimalFrequencyRatio}{See the momel source code for information on
this parameter.}

\item{intsint.level}{The name of the transcription level that should be
created and used for storing annotations. If this level already exists, it
will be used if attribute definitions match the ones given here.}

\item{attribute.definitions}{The names of attribute definitions to set up for
the transcription level. Currently not used.}

\item{absolute.tone.labels}{The INTSINT labels that should be set up as
members of the 'Absolute_tones' label group.}

\item{relative.tone.labels}{The INTSINT labels that should be set up as
members of the 'Relative_tones' label group.}

\item{force}{Force insertion of transcriptons. Currently not implemented.}

\item{verbose}{Should output be displayed for all major steps?}

\item{praat_path}{An explicit path to the Praat executable. Usually not
required.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} containing computed INTSINT (and MOMEL) information for
each bundle, with \code{start} time relative to the start of the bundle media
file.
}
\description{
This function transcribes the utterances specified in a segment list using
the MOMEL/INTSINT framework \insertCite{Hirst:2005jg,Hirst:2011wq}{reindeer} and insert
the EVENT labels into a special tier ("Intsint" by default). The function
will then call the Praat routines \insertCite{Hirst:2007ty}{reindeer} and the
external 'momel' program to make the calculations, and insert the resulting
levels onto the transcription tier. Please note that calling multiple
external programs results in some overhead, so please allow some time to
complete the transcription for large sets of utterances.
}
\details{
As a side effect, this function also sets up two label groups on the
attribute definition in which INTSINT labels are inserted (default
"Intsint"):
\enumerate{
\item Absolute_tones (T,M,B), and 2. Relative_tones (H,U,S,D,L)
}

which may be used in queries. In order to use them in larger queries, the
user has to define links between the created level and existing transcription
levels using \link{add_linkDefinition}, and links between created labels
on the two tiers using \link{autobuild_linkFromTimes}.

At the start of the MOMEL procedure, Pitch will computed using a wide
frequency search. The user may guide the search by supplying more restrictive
maximum and minimum fundamental frequency values (minF, maxF). By default,
the search considers a moderate pitch span (pitchSpan=1.5), but if very
expressive speech is investigated, \code{pitchSpan} may be increased to 2.5.

The MOMEL / INTSINT procedure also returns the MOMEL deduced f0  target
associated with the INTSINT label, as well as the f0 predicted by the INTSINT
model. These are returned by the function but currently not inserted as
labels in the transcriptions.
}
\examples{
\dontrun{
reindeer:::unlink_emuRDemoDir()
reindeer:::create_ae_db(verbose = TRUE) -> emuDBhandle
add_ssffTrackDefinition(emuDBhandle,"f0",onTheFlyFunctionName = "ksvF0")
query(emuDBhandle,"Intonational =~ .*") -> sl
annotate_INTSINT_MOMEL(emuDBhandle,sl) -> momelTab
add_perspective(emuDBhandle, "INTSINT")
set_signalCanvasesOrder(emuDBhandle,"INTSINT",c("OSCI","SPEC","f0"))
set_levelCanvasesOrder(emuDBhandle,"INTSINT",c("Intsint"))
serve(emuDBhandle)
}
}
\references{
\insertAllCited{}
}
