% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_annotate.R
\name{annotate_periods}
\alias{annotate_periods}
\title{Mark periods as events}
\usage{
annotate_periods(
  emuDBhandle,
  beginTime = 0,
  endTime = 0,
  windowShift = 5,
  minF = 75,
  maxF = 600,
  windowShape = "Gaussian1",
  interpolation = "cubic",
  RelativeWidth = 1,
  levelname = "AUTO_PERIODS",
  clear = TRUE
)
}
\arguments{
\item{emuDBhandle}{A emuR database handle.}

\item{beginTime}{Start time in seconds (0 for beginning of file)}

\item{endTime}{End time in seconds (0 for end of file)}

\item{windowShape}{Window shape for time windowing}

\item{interpolation}{The method used for finding intensity associated with the pitch period event (ont of "nearest", or "linear", "cubic", "sinc70" or "sinc700" interpolation).}

\item{RelativeWidth}{The relative with of the window used for extracting part of the signal for analysis.}

\item{levelname}{The name of the (EVENT) annotation level to insert into the database.}

\item{clear}{Should the \code{levename} level be cleared before inserting annotations?}
}
\value{
A \link[tibble:tibble]{tibble::tibble} containing information on the labels inserted (session, bundle, start, level, attribute, labels)
}
\description{
This function uses Praat \insertCite{Praatdoingphoneti:2022}{reindeer} to
extract a periodicity contour, and the associated intensity. The estimated
timing of each glottal cycle is marked at the maximum peak, and the timing
and intensity of that peak is then marked as an event with associated label in the
\code{levelname} transcription level.
}
\details{
If an intensity level could not be determined at the peak, the label of the
EVENT will be empty.
}
\references{
\insertAllCited{}
}
