% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_lazy_segment_list.R
\name{as.data.frame.lazy_segment_list}
\alias{as.data.frame.lazy_segment_list}
\title{Convert Lazy Segment List to Data Frame}
\usage{
\method{as.data.frame}{lazy_segment_list}(x, ...)
}
\arguments{
\item{x}{A lazy_segment_list object}

\item{...}{Additional arguments (ignored)}
}
\value{
A data.frame
}
\description{
Forces materialization
}
