% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_sequence_ops_optimized.R
\name{ascend_dt}
\alias{ascend_dt}
\title{Ascend Implementation Using data.table}
\usage{
ascend_dt(.segments, level, .from = NULL, .quiet = TRUE)
}
\description{
Ascend Implementation Using data.table
}
\keyword{internal}
