% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_transcription_system_optimized.R
\name{assess_optimized}
\alias{assess_optimized}
\title{Optimized assess() using data.table and batch operations}
\usage{
assess_optimized(suggestion, ...)
}
\description{
Optimized assess() using data.table and batch operations
}
