% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_executor.R
\name{build_simple_query_sql}
\alias{build_simple_query_sql}
\title{Build SQL for Simple Query}
\usage{
build_simple_query_sql(db_path, parsed)
}
\description{
Build SQL for Simple Query
}
\keyword{internal}
