% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_creation.R
\name{bundle_exists}
\alias{bundle_exists}
\title{Check if a bundle exists in session}
\usage{
bundle_exists(corpus_obj, session_name, bundle_name)
}
\arguments{
\item{corpus_obj}{A corpus object}

\item{session_name}{Session name}

\item{bundle_name}{Bundle name to check}
}
\value{
Logical; TRUE if bundle exists
}
\description{
Check if a bundle exists in session
}
\keyword{internal}
