% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_size_management.R
\name{clean_quantify_cache}
\alias{clean_quantify_cache}
\title{Clean quantify/enrich cache}
\usage{
clean_quantify_cache(
  corpus_obj,
  days_old = 30,
  dry_run = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{corpus_obj}{A corpus object}

\item{days_old}{Remove files older than this many days (default: 30)}

\item{dry_run}{If TRUE, only show what would be deleted}

\item{verbose}{Show messages}
}
\value{
Number of files deleted
}
\description{
Clean quantify/enrich cache
}
\keyword{internal}
