% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_size_management.R
\name{clean_simulation_cache}
\alias{clean_simulation_cache}
\title{Clean simulation cache}
\usage{
clean_simulation_cache(
  simulation_store,
  days_old = 30,
  dry_run = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{simulation_store}{Simulation store directory}

\item{days_old}{Remove files older than this many days (default: 30)}

\item{dry_run}{If TRUE, only show what would be deleted}

\item{verbose}{Show messages}
}
\value{
Number of files deleted
}
\description{
Clean simulation cache
}
\keyword{internal}
