% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_methods.R
\name{close_connection}
\alias{close_connection}
\title{Close the cached connection}
\usage{
close_connection(corpus_obj)
}
\arguments{
\item{corpus_obj}{A corpus object}
}
\description{
Close the cached connection
}
\keyword{internal}
