% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_lazy_segment_list.R
\name{collect}
\alias{collect}
\title{Collect/Materialize a Lazy Segment List}
\usage{
collect(lazy_sl, verbose = FALSE)
}
\arguments{
\item{lazy_sl}{A lazy_segment_list object}

\item{verbose}{Logical, whether to print SQL query}
}
\value{
A segment_list object
}
\description{
Forces execution of the query chain and returns a regular segment_list.
}
