% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_cmdi.R
\name{collect_database_metadata}
\alias{collect_database_metadata}
\title{Collect metadata from database structure}
\usage{
collect_database_metadata(db_handle, db_config, verbose = TRUE)
}
\description{
Collect metadata from database structure
}
\keyword{internal}
