% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_cmdi.R
\name{collect_participant_metadata}
\alias{collect_participant_metadata}
\title{Collect participant metadata from .meta_json files}
\usage{
collect_participant_metadata(db_handle, verbose = TRUE)
}
\description{
Collect participant metadata from .meta_json files
}
\keyword{internal}
