% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer-corpus.R
\name{corpus_assign_metadata}
\alias{corpus_assign_metadata}
\title{Assign metadata to corpus bundles/sessions}
\usage{
corpus_assign_metadata(
  corpus_obj,
  session_pattern,
  bundle_pattern,
  metadata_list
)
}
\description{
Assign metadata to corpus bundles/sessions
}
\keyword{internal}
