% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer-corpus.R
\name{corpus_import_media}
\alias{corpus_import_media}
\title{Import media files to corpus bundles}
\usage{
corpus_import_media(corpus_obj, session_pattern, bundle_pattern, media_spec)
}
\description{
Import media files to corpus bundles
}
\keyword{internal}
