% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_emuR_re-export.R
\name{create_links}
\alias{create_links}
\title{Create links between items. (A copy of the emuR function)}
\usage{
create_links(emuDBhandle, links, rewriteAllAnnots = TRUE, verbose = TRUE)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link[emuR]{load_emuDB}}}

\item{links}{data.frame like object containing linking information. The required columns
are: 
\itemize{
\item \code{session}: 
\item \code{bundle}
\item \code{from_id}
\item \code{to_id}
}}

\item{rewriteAllAnnots}{should changes be written to file system (_annot.json
files) (intended for expert use only)}

\item{verbose}{if set to \code{TRUE}, more status messages are printed}
}
\description{
create links between items
}
\seealso{
\link[emuR:create_links]{emuR::create_links}
}
