% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_sequence_ops_optimized.R
\name{descend_dt}
\alias{descend_dt}
\title{Descend Implementation Using data.table}
\usage{
descend_dt(.segments, level, .from = NULL, .quiet = TRUE)
}
\description{
Descend Implementation Using data.table
}
\keyword{internal}
