% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_core.R
\name{deserialize_metadata_value}
\alias{deserialize_metadata_value}
\title{Deserialize a metadata value from storage}
\usage{
deserialize_metadata_value(value_str, type_str)
}
\description{
Deserialize a metadata value from storage
}
\keyword{internal}
