% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_autosync.R
\name{detect_annot_changes}
\alias{detect_annot_changes}
\title{Detect changes in _annot.json files}
\usage{
detect_annot_changes(db_handle)
}
\arguments{
\item{db_handle}{An emuDBhandle object}
}
\value{
Data frame with changed bundles (session, bundle, annot_path)
}
\description{
Detect changes in _annot.json files
}
\keyword{internal}
