% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_autosync.R
\name{detect_config_changes}
\alias{detect_config_changes}
\title{Detect changes in database configuration}
\usage{
detect_config_changes(db_handle)
}
\arguments{
\item{db_handle}{An emuDBhandle object}
}
\value{
Logical; TRUE if configuration changed
}
\description{
Detect changes in database configuration
}
\keyword{internal}
