% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_autosync.R
\name{detect_metadata_changes}
\alias{detect_metadata_changes}
\title{Detect changes in .meta_json files}
\usage{
detect_metadata_changes(db_handle)
}
\arguments{
\item{db_handle}{An emuDBhandle object}
}
\value{
Data frame with changed metadata files
}
\description{
Detect changes in .meta_json files
}
\keyword{internal}
