% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_io.R
\name{.add_bundle_to_db}
\alias{.add_bundle_to_db}
\title{Add bundle record to SQLite}
\usage{
.add_bundle_to_db(con, db_uuid, session, name, annotates, sample_rate, md5)
}
\arguments{
\item{con}{DBI connection}

\item{db_uuid}{Database UUID}

\item{session}{Session name}

\item{name}{Bundle name}

\item{annotates}{What the bundle annotates (e.g. wav file)}

\item{sample_rate}{Sample rate}

\item{md5}{MD5 hash of annotation JSON}
}
\description{
Inserts a new bundle record into the bundle table of the emuDBcache.sqlite.
}
\keyword{internal}
