% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_serve.R
\docType{data}
\name{.emu_suffixes}
\alias{.emu_suffixes}
\title{EMU file suffixes}
\format{
An object of class \code{list} of length 3.
}
\usage{
.emu_suffixes
}
\description{
Constants for EMU database file naming conventions.
These replicate get_session_suffix(), get_bundle_dir_suffix(), etc.
}
\note{
Replaces emuR internal constants to avoid ::: dependency
}
\keyword{internal}
