% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_listing.R
\name{.extract_basePath_dbName}
\alias{.extract_basePath_dbName}
\title{Extract basePath and dbName from a corpus or emuDBhandle}
\usage{
.extract_basePath_dbName(obj)
}
\arguments{
\item{obj}{A corpus object or emuDBhandle-like list}
}
\value{
A list with \code{basePath} and \code{dbName}
}
\description{
Extract basePath and dbName from a corpus or emuDBhandle
}
\keyword{internal}
