% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_import_export.R
\name{.is_metadata_dirty}
\alias{.is_metadata_dirty}
\title{Test whether a corpus has pending metadata changes}
\usage{
.is_metadata_dirty(corpus_obj)
}
\description{
Test whether a corpus has pending metadata changes
}
\keyword{internal}
