% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_methods.R
\name{.subset_corpus}
\alias{.subset_corpus}
\title{Subset corpus to get bundle list with metadata}
\usage{
.subset_corpus(x, i, j, ..., drop = FALSE)
}
\description{
Subset corpus to get bundle list with metadata
}
\keyword{internal}
