% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_io.R
\name{.write_bundle_list}
\alias{.write_bundle_list}
\title{Write a bundle list file}
\usage{
.write_bundle_list(basePath, name, data)
}
\arguments{
\item{basePath}{Database basePath}

\item{name}{Bundle list name}

\item{data}{tibble/data.frame with bundle list data}
}
\value{
Invisible path to written file
}
\description{
Writes bundle list data to a JSON file in the database basePath.
}
\keyword{internal}
