% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_cache_system.R
\name{draft_cache_summary}
\alias{draft_cache_summary}
\title{Get draft cache summary}
\usage{
draft_cache_summary(corpus_obj, draft_function_name, date = Sys.Date())
}
\arguments{
\item{corpus_obj}{A corpus object}

\item{draft_function_name}{Name of draft function}

\item{date}{Date of cache (defaults to today)}
}
\value{
List with cache statistics
}
\description{
Get draft cache summary
}
