% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_signalextensions_dt.R
\name{dspp_metadataParameters_dt}
\alias{dspp_metadataParameters_dt}
\title{Optimized version of dspp_metadataParameters using data.table}
\usage{
dspp_metadataParameters_dt(
  recompute = FALSE,
  id.columns = c("Age", "Gender"),
  impute = TRUE,
  defaultsEstimatedSampleSize = 10
)
}
\arguments{
\item{recompute}{If \code{TRUE}, this function will estimate age appropriate
settings from the collection of empirical defaults stored in this package.
If \code{FALSE}, this function will just return the stored \code{DSPP} data.}

\item{id.columns}{Identification columns (default: c("Age","Gender"))}

\item{impute}{Impute missing settings using time series analysis?}

\item{defaultsEstimatedSampleSize}{The estimated sample size on which software
default settings have been based.}
}
\value{
A data.table with one row per age and gender combination
}
\description{
This function returns age and gender specific speech signal processing
parameters with significantly improved performance using data.table.
}
