% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_psychoacoustics.R
\name{erb}
\alias{erb}
\title{Convert Herz to Semitones Equivalent Rectangular Bandwidth (ERB)}
\usage{
erb(f)
}
\arguments{
\item{f}{A vector of frequency values (in Hz)}
}
\value{
A vector of ERB values
}
\description{
Convert physical frequencies (in Hz) to number of Equivalent rectangular
bandwidth (ERB) below the frequency. The ERB is  an approximation computed
from the bandwidths of filters in human hearing, as modeled by rectangular
band-pass filters.
}
\details{
The formula for Number of ERBs given in
\insertCite{Moore:1982ha}{reindeer} is used
in the calculations.
}
\examples{
f <- seq(1,10000,10)
plot(f,erb(f))
plot(erb(f),emuR::bark(f),lty=2,ylab="Zwicker et al. Bark scale")
abline(a = 0, b=1,col="blue")
}
\references{
\insertAllCited{}
}
