% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer-corpus.R
\name{gather_metadata_internal}
\alias{gather_metadata_internal}
\title{Gather all metadata from .meta_json files (internal, called during construction)}
\usage{
gather_metadata_internal(corpus_obj, verbose = FALSE)
}
\description{
Gather all metadata from .meta_json files (internal, called during construction)
}
\keyword{internal}
