% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_cmdi.R
\name{generate_cmdi_xml}
\alias{generate_cmdi_xml}
\title{Generate CMDI XML document}
\usage{
generate_cmdi_xml(
  profile,
  db_name,
  db_uuid,
  corpus_title,
  corpus_description,
  author,
  institution,
  contact_email,
  license,
  availability,
  db_metadata,
  participant_metadata,
  include_placeholders
)
}
\description{
Generate CMDI XML document
}
\keyword{internal}
