% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_cache_system.R
\name{get_draft_cache_path}
\alias{get_draft_cache_path}
\title{Get full path to draft cache file}
\usage{
get_draft_cache_path(corpus_obj, draft_function_name, date = Sys.Date())
}
\arguments{
\item{corpus_obj}{A corpus object}

\item{draft_function_name}{Name of the draft function}

\item{date}{Date for cache file (defaults to today)}
}
\value{
Full path to cache file
}
\description{
Get full path to draft cache file
}
\keyword{internal}
