% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_size_management.R
\name{get_file_size}
\alias{get_file_size}
\title{Get file size}
\usage{
get_file_size(path, recursive = FALSE)
}
\arguments{
\item{path}{Path to file or directory}

\item{recursive}{If TRUE and path is directory, get total size recursively}
}
\value{
Size in bytes
}
\description{
Get file size
}
\keyword{internal}
