% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_emuR_re-export.R
\name{get_levelCanvasesOrder}
\alias{get_levelCanvasesOrder}
\title{Get the level order of a perspective. (A copy of the emuR function)}
\usage{
get_levelCanvasesOrder(emuDBhandle, perspectiveName)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link[emuR]{load_emuDB}}}

\item{perspectiveName}{name of perspective}
}
\description{
Set / Get which levels of an emuDB to display as level canvases (in a 
given perspective of the EMU-webApp),
and in what order. Level canvases refer to levels of 
the type "SEGMENT" or "EVENT" that are displayed by the EMU-webApp. Levels 
of type "ITEM" can always be displayed using the hierarchy view of the
web application but can not be displayed as level canvases.
For more information on the structural elements of an emuDB 
see \code{vignette{emuDB}}.
}
\seealso{
\link[emuR:SetGetlevelCanvasesOrder]{emuR::get_levelCanvasesOrder}
}
