% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_transcription_system.R
\name{get_levelDefinition_from_config}
\alias{get_levelDefinition_from_config}
\title{Get level definition from config}
\usage{
get_levelDefinition_from_config(config, level_name)
}
\description{
Get level definition from config
}
\keyword{internal}
