% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_metadata_optimized.R
\name{get_metadata_diagnostics_internal}
\alias{get_metadata_diagnostics_internal}
\title{Get metadata diagnostics}
\usage{
get_metadata_diagnostics_internal(con, db_uuid)
}
\description{
Get metadata diagnostics
}
\keyword{internal}
