% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_metadata_optimized.R
\name{get_metadata_field}
\alias{get_metadata_field}
\title{Get values for a single metadata field across bundles}
\usage{
get_metadata_field(con, db_uuid, field_name, sessions, bundles)
}
\description{
Get values for a single metadata field across bundles
}
\keyword{internal}
