% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer-corpus.R
\name{get_metadata_field_values}
\alias{get_metadata_field_values}
\title{Get values for a metadata field with inheritance}
\usage{
get_metadata_field_values(con, db_uuid, field_name, sessions, bundles)
}
\description{
Get values for a metadata field with inheritance
}
\keyword{internal}
