% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer-corpus.R
\name{get_metadata_for_patterns}
\alias{get_metadata_for_patterns}
\title{Get metadata for bundles matching patterns}
\usage{
get_metadata_for_patterns(corpus_obj, session_pattern, bundle_pattern)
}
\description{
Get metadata for bundles matching patterns
}
\keyword{internal}
