% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_signalextensions.R
\name{get_metaFuncFormals}
\alias{get_metaFuncFormals}
\title{A utility function that builds a list of parameters to use in function call based on metadata}
\usage{
get_metaFuncFormals(
  emuDBhandle,
  session,
  bundle,
  onTheFlyFunctionName,
  onTheFlyParams = list(),
  metadata.defaults = list(Gender = NA, Age = 35),
  package = "superassp"
)
}
\arguments{
\item{session}{Only consider this session.}

\item{bundle}{The bundle in the session to build an argument list for.}
}
\value{
A list which may be supplied to a \code{do.call} call.
}
\description{
This function takes a function and explicit argument list and builds a list of parameters for a function call considering the metadata of a bundle (in a session).
}
\seealso{
match_parameters

do.call
}
