% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_size_management.R
\name{get_quantify_cache_dir}
\alias{get_quantify_cache_dir}
\title{Get cache directory path for quantify}
\usage{
get_quantify_cache_dir(corpus_obj)
}
\arguments{
\item{corpus_obj}{A corpus object}
}
\value{
Path to cache directory
}
\description{
Get cache directory path for quantify
}
\keyword{internal}
