% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_signalextensions.R
\name{get_ssffObject}
\alias{get_ssffObject}
\title{Allows the user to get an SSFF object directly from an Emu database}
\usage{
get_ssffObject(emuDBhandle, extension, n)
}
\arguments{
\item{emuDBhandle}{An Emu database handle.}

\item{extension}{A file extension.}

\item{n}{The index of the track file to be returned as an SSFF object.}
}
\value{
An SSFF track object
}
\description{
The user gives a database handle, a file extension of files that are present
in the Emu database directory (i.e. have been generated using
\link[emuR:AddListRemoveSsffTrackDefinition]{emuR::add_ssffTrackDefinition} or \link{add_trackDefinition}) calls),
an index and get the 'n'th track file with the file extension in the database
as an SSFF object.
}
