% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_signalextensions.R
\name{get_trackColumns}
\alias{get_trackColumns}
\title{Get the columns / fields defined in SSFF tracks with a given extension}
\usage{
get_trackColumns(emuDBhandle, extension = NULL)
}
\arguments{
\item{emuDBhandle}{The EmuR database handle.}

\item{extension}{The file extension of the SSFF files to investigated.}
}
\value{
A vector containing SSFF field / column names within the SSFF file.
}
\description{
This function allows the user to specify an EmuR database handle and a
file extension and get which columns or tracks are defined in the SSFF files.
}
\examples{
\dontrun{
 ae <- reindeer:::create_ae_db()
 reindeer::add_trackDefinition(ae,"A","A",onTheFlyFunctionName = "praat_sauce")
 print(reindeer::get_trackColumns(ae,"psa"))
}
}
