% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_serve.R
\name{get_webapp_dir}
\alias{get_webapp_dir}
\title{Get EMU-webApp directory path}
\usage{
get_webapp_dir()
}
\value{
Path to EMU-webApp dist directory
}
\description{
Returns the path to the revised EMU-webApp distribution directory.
The function looks for the EMU-webApp in this order:
\enumerate{
\item Option: getOption("reindeer.emuWebApp.dir")
\item Environment variable: EMU_WEBAPP_DIR
\item Package installation: system.file("EMU-webApp/dist", package = "reindeer")
\item Default fallback: ../EMU-webApp/dist relative to package location
}
}
\keyword{internal}
