% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_segment_list.R
\name{glimpse.extended_segment_list}
\alias{glimpse.extended_segment_list}
\title{Glimpse method for extended_segment_list}
\usage{
\method{glimpse}{extended_segment_list}(x, width = NULL, ...)
}
\description{
Glimpse method for extended_segment_list
}
