% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_lazy_segment_list.R
\name{glimpse.lazy_segment_list}
\alias{glimpse.lazy_segment_list}
\title{Glimpse method for lazy_segment_list}
\usage{
\method{glimpse}{lazy_segment_list}(x, ...)
}
\description{
Glimpse method for lazy_segment_list
}
