% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_segment_list.R
\name{glimpse.segment_list}
\alias{glimpse.segment_list}
\title{Glimpse method for segment_list}
\usage{
\method{glimpse}{segment_list}(x, width = NULL, ...)
}
\description{
Glimpse method for segment_list
}
