% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer.R
\name{\%||\%}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Value to check}

\item{y}{Fallback value if x is NULL or length-0}
}
\description{
Null coalescing operator
}
\keyword{internal}
