% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_serve.R
\name{guess_mime_type}
\alias{guess_mime_type}
\title{Guess MIME type from file extension}
\usage{
guess_mime_type(path)
}
\arguments{
\item{path}{File path}
}
\value{
MIME type string
}
\description{
Replicates emuR:::guess_type for common file types used in EMU.
}
\note{
Replaces emuR:::guess_type to avoid internal dependency
}
\keyword{internal}
