% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_annotate_python.R
\name{has_python_annotations}
\alias{has_python_annotations}
\title{Check if Python annotation environment is available}
\usage{
has_python_annotations()
}
\description{
Check if Python annotation environment is available
}
