% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer-corpus.R
\name{import_media_to_bundle}
\alias{import_media_to_bundle}
\title{Import media file to a specific bundle}
\usage{
import_media_to_bundle(
  corpus_obj,
  session_name,
  bundle_name,
  file_path,
  channel_map
)
}
\description{
Import media file to a specific bundle
}
\keyword{internal}
