% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_database.R
\name{import_recordings}
\alias{import_recordings}
\title{Import a speech recordings into an Emu database session}
\usage{
import_recordings(
  emuDBhandle,
  dir,
  media_pattern = ".*",
  targetSessionName = "0000",
  downsample.to = NULL,
  verbose = TRUE
)
}
\arguments{
\item{emuDBhandle}{The Emu database handle.}

\item{dir}{The directory containing speech recordings or session directories.}

\item{targetSessionName}{The default session name, if not specified using sub-directories in the import folder.}

\item{downsample.to}{Set the sampling frequency to downsample the input file to before writing to the database. A \code{NULL} indicates
that the original sampling frequency of the file will be left intact.}

\item{verbose}{Should additional information and progress bar be displayed to the user?}
}
\description{
This function imports speech recordings stored in a directory into session in
an Emu database. The new session will be named according to the name the user
provides using the \code{targetSessionName} argument. A new bundle will be created
for every file with the file extension specified in the "mediafileExtension"
setting in the database configuration file (usually "wav"). ALternatively,
the user may set up sub-directories in the import folder, which will then be
used as recording session names.
}
\details{
In order to ensure that the database always contains a mono sound file, the
single channel indicated by the \code{speech.channel} argument will be extracted
and stored in the database.

The function also makes sure that 'OSCI' and 'SPEC' perspectives are defined
for the database.
}
