% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_annotate_python.R
\name{init_annotation_python}
\alias{init_annotation_python}
\title{Initialize Python environment for annotations}
\usage{
init_annotation_python()
}
\description{
Initialize Python environment for annotations
}
\keyword{internal}
