% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_autosync.R
\name{init_sync_state}
\alias{init_sync_state}
\title{Initialize synchronization state tracking}
\usage{
init_sync_state(db_handle)
}
\arguments{
\item{db_handle}{An emuDBhandle object}
}
\description{
Initialize synchronization state tracking
}
\keyword{internal}
