% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_cache_system.R
\name{initialize_draft_cache}
\alias{initialize_draft_cache}
\title{Initialize draft cache database}
\usage{
initialize_draft_cache(cache_path, draft_function_name)
}
\arguments{
\item{cache_path}{Path to cache file}

\item{draft_function_name}{Name of the draft function}
}
\value{
Connection to cache database
}
\description{
Creates SQLite database for storing draft annotations with metadata.
}
\keyword{internal}
