% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_core.R
\name{initialize_metadata_schema}
\alias{initialize_metadata_schema}
\title{Initialize metadata tables in SQLite cache}
\usage{
initialize_metadata_schema(con)
}
\arguments{
\item{con}{SQLite connection}
}
\description{
Adds metadata-specific tables to the cache database
}
\keyword{internal}
