% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_simulation.R
\name{initialize_simulation_cache}
\alias{initialize_simulation_cache}
\title{Initialize simulation cache database}
\usage{
initialize_simulation_cache(cache_dir, timestamp, dsp_function_name)
}
\arguments{
\item{cache_dir}{Directory for simulation cache}

\item{timestamp}{Timestamp for cache file name}

\item{dsp_function_name}{Name of DSP function}
}
\value{
Path to created cache file
}
\description{
Creates SQLite database to store simulation results
}
\keyword{internal}
