% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_cache_system.R
\name{is_bundle_cached}
\alias{is_bundle_cached}
\title{Check if bundle is already in cache}
\usage{
is_bundle_cached(con, session, bundle, level_name, attribute_name)
}
\arguments{
\item{con}{Database connection}

\item{session}{Session name}

\item{bundle}{Bundle name}

\item{level_name}{Level name}

\item{attribute_name}{Attribute name}
}
\value{
Logical; TRUE if bundle is cached
}
\description{
Check if bundle is already in cache
}
\keyword{internal}
