% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_lazy_segment_list.R
\name{is_lazy}
\alias{is_lazy}
\title{Check if Object is Lazy}
\usage{
is_lazy(x)
}
\arguments{
\item{x}{An object}
}
\value{
Logical
}
\description{
Check if Object is Lazy
}
\keyword{internal}
