% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_list_classes.R
\name{is_segment_list}
\alias{is_segment_list}
\title{Check if object is a segment_list}
\usage{
is_segment_list(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical
}
\description{
Check if object is a segment_list
}
