% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_emuR_re-export.R
\name{list_bundles}
\alias{list_bundles}
\title{List sessions in an emuDB. (A copy of the emuR function)}
\usage{
list_bundles(emuDBhandle, session = NULL)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link[emuR]{load_emuDB}}}

\item{session}{optional session}
}
\value{
data.frame object with columns session and name of bundles
}
\description{
List all bundles of emuDB or of particular session.
}
\seealso{
\link[emuR:list_bundles]{emuR::list_bundles}
}
