% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_cache_system.R
\name{list_draft_caches}
\alias{list_draft_caches}
\title{List all draft caches for a corpus}
\usage{
list_draft_caches(corpus_obj)
}
\arguments{
\item{corpus_obj}{A corpus object}
}
\value{
Data frame with cache information
}
\description{
List all draft caches for a corpus
}
