% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_emuR_re-export.R
\name{list_files}
\alias{list_files}
\title{Re-exported function from emuR list the files of an emuDB. (A copy of the emuR function)}
\usage{
list_files(
  emuDBhandle,
  fileExtension = ".*",
  sessionPattern = ".*",
  bundlePattern = ".*"
)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link[emuR]{load_emuDB}}}

\item{fileExtension}{file extension of files}

\item{sessionPattern}{A (RegEx) pattern matching sessions to be searched from the database}

\item{bundlePattern}{A (RegEx) pattern matching bundles to be searched from the database}
}
\value{
file paths as character vector
}
\description{
List files belonging to emuDB. For 
more information on the structural elements of an emuDB 
see \code{vignette{emuDB}}.
}
\seealso{
\link[emuR:list_files]{emuR::list_files}
}
