% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_emuR_re-export.R
\name{list_linkDefinitions}
\alias{list_linkDefinitions}
\title{List level definitions. (A copy of the emuR function)}
\usage{
list_linkDefinitions(emuDBhandle)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link[emuR]{load_emuDB}}}
}
\description{
Add / List / Remove new link definition to / of / from emuDB. A link definition
specifies the relationship between two levels, the
super-level and the sub-level. The entirety of all link 
definitions of a emuDB specifies the 
hierarchical structure of the database. For more information
on the structural elements of an emuDB see \code{vignette(emuDB)}.
}
\details{
Link type descriptions:
\itemize{
\item{\code{"ONE_TO_MANY"}}{A single ITEM of the super-level can be linked to multiple ITEMs of the sub-level}
\item{\code{"MANY_TO_MANY"}}{Multiple ITEMs of the super-level can be linked to multiple ITEMs of the sub-level}
\item{\code{"ONE_TO_ONE"}}{A single ITEM of the super-level can be linked to a single ITEM of the sub-level}
}

For all link types the rule applies that no links are allowed to cross any other links.
Further, a linkDefinition can not be removed, if there are links present in the emuDB.
}
\seealso{
\link[emuR:AddListRemoveLinkDefinition]{emuR::list_linkDefinitions}
}
