% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_emuR_re-export.R
\name{list_sampleRates}
\alias{list_sampleRates}
\title{List sample rates of media and annotation (_annot.json) files (from emuR)}
\usage{
list_sampleRates(emuDBhandle, sessionPattern = ".*", bundlePattern = ".*")
}
\arguments{
\item{emuDBhandle}{emuDB handle object (see \link[emuR]{load_emuDB})}

\item{sessionPattern}{A regular expression pattern matching session names to be searched from the database}

\item{bundlePattern}{A regular expression pattern matching bundle names to be searched from the database}
}
\value{
tibble with the columns 
\itemize{
\item session
\item bundle
\item sample_rate_annot_json
\item sample_rate_media_file
}
\code{session}, \code{b}
}
\description{
List sample rates of media and annotation (_annot.json) files
}
\seealso{
\link[emuR:list_sampleRates]{emuR::list_sampleRates}
}
