% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_emuR_re-export.R
\name{list_sessions}
\alias{list_sessions}
\title{Re-exported function from emuR list sessions in an emuDB (from emuR)}
\usage{
list_sessions(emuDBhandle)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link[emuR]{load_emuDB}}}
}
\value{
data.frame object with session names
}
\description{
List session names of emuDB
}
\seealso{
\link[emuR:list_sessions]{emuR::list_sessions}
}
