% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_metadata_optimized.R
\name{list_sessions_from_cache}
\alias{list_sessions_from_cache}
\title{List sessions from cache}
\usage{
list_sessions_from_cache(con, db_uuid)
}
\description{
List sessions from cache
}
\keyword{internal}
