% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribe.R
\name{list_untranscribedBundles}
\alias{list_untranscribedBundles}
\title{Find untranscribed bundles}
\usage{
list_untranscribedBundles(emuDBhandle)
}
\arguments{
\item{emuDBhandle}{}
}
\value{
A vector of bundle names
}
\description{
Find untranscribed bundles
}
\examples{
\\donotrun{
ae <- create_ae_db()
# Make a test case for this particular function
currFile <- file.path(ae$basePath,"0000_ses","msajc012_bndl","msajc012_annot.json")
unlink(currFile,force=TRUE)
print(list_untranscribedBundles(ae))
unlink_emuRDemoDir()
}

}
\author{
Fredrik Karlsson
}
