% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_emuR_re-export.R
\name{load_emuDB}
\alias{load_emuDB}
\title{Load an emuDB (from emuR)}
\usage{
load_emuDB(
  emuDBhandle,
  sessionPattern = ".*",
  bundlePattern = ".*",
  seglist = NULL,
  bundleListName = NULL,
  host = "127.0.0.1",
  port = 17890,
  autoOpenURL = "https://ips-lmu.github.io/EMU-webApp/?autoConnect=true",
  browser = getOption("browser"),
  useViewer = TRUE,
  debug = FALSE,
  debugLevel = 0
)

load_emuDB(
  emuDBhandle,
  sessionPattern = ".*",
  bundlePattern = ".*",
  seglist = NULL,
  bundleListName = NULL,
  host = "127.0.0.1",
  port = 17890,
  autoOpenURL = "https://ips-lmu.github.io/EMU-webApp/?autoConnect=true",
  browser = getOption("browser"),
  useViewer = TRUE,
  debug = FALSE,
  debugLevel = 0
)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link[emuR]{load_emuDB}}}

\item{sessionPattern}{A regular expression pattern matching session names to be served}

\item{bundlePattern}{A regular expression pattern matching bundle names to be served}

\item{seglist}{segment list to use for times anchors and session + bundle restriction (type: \link[emuR]{emuRsegs})}

\item{bundleListName}{name of bundleList stored in emuDB/bundleLists subdir to send to EMU-webApp}

\item{host}{host IP to listen to (default: 127.0.0.1  (localhost))}

\item{port}{the port number to listen on (default: 17890)}

\item{autoOpenURL}{URL passed to \code{\link{browseURL}} function. If NULL or an empty string are passed in
\code{\link{browseURL}} will not be invoked.}

\item{browser}{argument passed on to \code{browser} argument of \code{\link{browseURL}} (see 
it's documentation for details )}

\item{useViewer}{Use the viewer provided by \code{getOption("viewer")} (the viewer pane when using RStudio) 
and host a local version of the EMU-webApp in it. This will clone the current 
EMU-webApp build (\url{https://github.com/IPS-LMU/EMU-webApp/tree/gh-pages/}) into the directory provided by 
\code{\link{tempdir}} and serve this local version. A clone will 
only be performed if no \code{file.path(tempdir(), "EMU-webApp")} directory is present. An alternative
directory can be also set: \code{options(emuR.emuWebApp.dir="path/to/EMU-webApp")} (use if offline functionality is required).}

\item{debug}{TRUE to enable debugging (default: no debugging messages)}

\item{debugLevel}{integer higher values generate more detailed debug output}
}
\value{
emuDB handle object
}
\description{
Function loads emuDB into its cached representation and makes it accessible from within the 
current R session by returning a emuDBhandle object
}
\details{
In order to access an emuDB from R it is necessary to load the annotation and configuration 
files to an emuR internal database format. The function expects a emuDB file structure in directory 
\code{databaseDir}. The emuDB configuration file is loaded first. On success the function iterates 
through session and bundle directories and loads found annotation files. The parameter \code{inMemoryCache} 
determines where the internal database is stored: If \code{FALSE} a database cache file in \code{databaseDir} 
is used. When the database is loaded for the first time the function will create a new cache file and store 
the data to it. On subsequent loading of the same database the cache is only updated if files have changed, 
therefore the loading is then much faster. For this to work the user needs write permissions to 
\code{databaseDir} and the cache file. The database is loaded into a volatile in-memory database if 
\code{inMemoryCache} is set to \code{TRUE}.
}
\seealso{
\link[emuR:load_emuDB]{emuR::load_emuDB}

\link[emuR:serve]{emuR::serve}
}
