% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_autosync.R
\name{load_sync_config}
\alias{load_sync_config}
\title{Load sync configuration}
\usage{
load_sync_config(db_handle)
}
\arguments{
\item{db_handle}{An emuDBhandle object}
}
\value{
List with sync configuration, or NULL if not configured
}
\description{
Load sync configuration
}
\keyword{internal}
