% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_autosync.R
\name{load_sync_config_from_path}
\alias{load_sync_config_from_path}
\title{Load sync config from database path (non-handle version)}
\usage{
load_sync_config_from_path(basePath)
}
\description{
Load sync config from database path (non-handle version)
}
\keyword{internal}
