% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_autosync.R
\name{load_sync_state}
\alias{load_sync_state}
\title{Load sync state}
\usage{
load_sync_state(db_handle)
}
\arguments{
\item{db_handle}{An emuDBhandle object}
}
\value{
List with sync state
}
\description{
Load sync state
}
\keyword{internal}
