% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_signalextensions.R
\name{metadata_parameters}
\alias{metadata_parameters}
\title{Metadata-aware selection of DSP parameters}
\usage{
metadata_parameters(
  emuDBhandle,
  onTheFlyFunctionName,
  onTheFlyParams = NULL,
  metadata.defaults = list(Gender = NA, Age = 35),
  package = "superassp",
  DSP.file = NULL,
  wide.format = TRUE
)
}
\arguments{
\item{emuDBhandle}{The database handle}

\item{onTheFlyFunctionName}{The name of the function for which appropriate
parameters should be deduced.}

\item{onTheFlyParams}{A list of explicit DSP function parameters.}

\item{metadata.defaults}{Default settings for metadata. This argument is
primarily used for setting default Age and Gender for bundles with no such
metadata already set.}

\item{package}{The signal processing package where the \code{onTheFlyFunctionName}
function may be found. Defaults to \link{superassp}.}

\item{DSP.file}{The full path of an Excel file containig gender and age
specific DSP parameters. Please consult \link{get_parameters} for more
information. If \code{NULL}, the default \code{DSPP} data frame will be used to get
DSP parameters.}

\item{wide.format}{If FALSE, the output will be in long format rather than
the default wide format.}
}
\value{
A \link{tibble} with columns \code{session}, \code{bundle}, and then one column per
parameter value for the \code{onTheFlyFunctionName} function. The output is
grouped by \code{session} and \code{bundle}.
}
\description{
This function deduces the appropriate parameters for a speech signal
processing function based on the speaker Age and Gender metadata set for a
bundle.
}
\details{
The deduction of parameters will be done in four steps:

1.The user can specify metadata values for cases where for instance Age and
Gender is not specified for a bundle. These will be injected into the stored
metadata before extracting DSP parameters.
\enumerate{
\item The function will use Age and Gender metadata for each bundle (or from the
session) and consults the \link{DSPP} set of parameters to deduce function
arguments. A Gender of NA means that an gender neutral, and therefore
possibly very wide, settings will be applied. The procedure will select the
appropriate age-dependent setting with the most specific (smallest) age
range. As a result, one can have a set of specifications like these:
{Gender=NA, Age_lower=0, Age_upper=200}, and then {Gender=NA, Age_lower=10,
Age_upper=20} and the procedure will be able to select the second set for any
15 year old person for which a better setting is not found. For a Male 15
year old speaker, these settings will also be applied, if there are no other
set that is defined for, for instance, {Gender=Male, Age_lower=10,
Age_upper=20}.
\item Missing function parameters for a bundle after having considered metadata
and the DSP parameter sets will be completed from the formal default
arguments of the \code{onTheFlyFunctionName} function.
\item If the user specifies a set of explicit function parameters in the
\code{onTheFlyParams}, these will overwrite any parameter values deduced from the
metadata.
}
}
\examples{
\dontrun{
reindeer:::unlink_emuRDemoDir()
reindeer:::create_ae_db() -> emuDBhandle
reindeer:::make_dummy_metafiles(emuDBhandle)
print(metadata_parameters(emuDBhandle,onTheFlyFunctionName = "forest") )
}
}
