% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_transcription_system_optimized.R
\name{optimize_suggestion_data}
\alias{optimize_suggestion_data}
\title{Convert Suggestion internal data to data.table}
\usage{
optimize_suggestion_data(suggestion)
}
\description{
Convert Suggestion internal data to data.table
}
\keyword{internal}
