% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_parser.R
\name{parse_eql_query}
\alias{parse_eql_query}
\title{Query EMU database using optimized SQLite backend}
\usage{
parse_eql_query(query_string)
}
\arguments{
\item{query_string}{EQL query string to parse}
}
\value{
A list describing the parsed query structure
}
\description{
Query EMU database using optimized SQLite backend
}
\keyword{internal}
