% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_list_pivot.R
\name{print.track_long}
\alias{print.track_long}
\title{Print method for track_long tibbles}
\usage{
\method{print}{track_long}(x, ...)
}
\arguments{
\item{x}{A \code{track_long} object.}

\item{...}{Passed to the tibble printer.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for track_long tibbles
}
\keyword{internal}
