% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_signalextensions_dt.R
\name{process_unspecified_data}
\alias{process_unspecified_data}
\title{Process unspecified gender data with data.table}
\usage{
process_unspecified_data(defaults, impute, defaultsEstimatedSampleSize)
}
\description{
Process unspecified gender data with data.table
}
\keyword{internal}
