% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_emuR_re-export.R
\name{read_bundleList}
\alias{read_bundleList}
\title{Read a stored bundleList. (A copy of the emuR function)}
\usage{
read_bundleList(emuDBhandle, name)
}
\arguments{
\item{emuDBhandle}{emuDB handle object (see \link[emuR]{load_emuDB})}

\item{name}{name of bundleList (excluding the _bundleList.json suffix)}
}
\value{
tibble with the columns \code{session}, \code{name}, 
\code{comment}, \code{finishedEditing}
}
\description{
read bundleList JSON file in emuDB
}
\details{
Read bundleList JSON file in emuDB that is stored in
the databases root dir sub-dir \code{bundleLists/}
}
\seealso{
\link[emuR:read_bundleList]{emuR::read_bundleList}
}
