% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR.R
\name{reindeer}
\alias{reindeer}
\title{reindeer: An extension of the capabilities of emuR to work with speech data in
a nordic climate}
\description{
This package extends the capabilities of emuR to work with speech  data and
includes the facilities for working with metadata, and for working with
computationally intensive and sparse measurement data. The package also
provides functions for setting up Emu web app with signal track overlays.When
loaded, it also loads many functions from the emuR package, and makes them
available for the user.
}
