% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_emuR_re-export.R
\name{remove_attrDefLabelGroup}
\alias{remove_attrDefLabelGroup}
\title{Remove a label group from an attribute definition (from emuR)}
\usage{
remove_attrDefLabelGroup(
  emuDBhandle,
  levelName,
  attributeDefinitionName,
  labelGroupName
)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link[emuR]{load_emuDB}}}

\item{levelName}{name of level}

\item{attributeDefinitionName}{name of attributeDefinition}

\item{labelGroupName}{name of label group}
}
\description{
Add / List / Remove label group to / of / from a specific attribute definition. 
This label group can be used as a short hand  
to reference groups of labels specific
to an attribute definition (compared to global label groups that 
are added by \code{\link[emuR]{add_labelGroup}}) in a 
\code{\link[emuR]{query}}. A common example would be to
add a label group for something like the phonetic
category of nasals to be able reference them 
as "nasals" in a \code{\link[emuR]{query}}. For more information 
on the structural elements of an emuDB see \code{vignette(emuDB)}.
}
\seealso{
\link[emuR:AddListRemoveAttrDefLabelGroup]{emuR::remove_attrDefLabelGroup}
}
