% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_size_management.R
\name{remove_old_cache_files}
\alias{remove_old_cache_files}
\title{Remove old cache files by age}
\usage{
remove_old_cache_files(
  cache_dir,
  days_old = 30,
  pattern = "\\\\.sqlite$",
  dry_run = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{cache_dir}{Cache directory}

\item{days_old}{Remove files older than this many days}

\item{pattern}{File pattern to match}

\item{dry_run}{If TRUE, only show what would be deleted}

\item{verbose}{Show messages}
}
\value{
Number of files deleted
}
\description{
Remove old cache files by age
}
\keyword{internal}
