% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_emuR_re-export.R
\name{remove_ssffTrackDefinition}
\alias{remove_ssffTrackDefinition}
\title{Remove an SSFF track from the emuDB. (A copy of the emuR function)}
\usage{
remove_ssffTrackDefinition(emuDBhandle, name, deleteFiles = FALSE)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link[emuR]{load_emuDB}}}

\item{name}{name of ssffTrackDefinition}

\item{deleteFiles}{delete files that belong to ssffTrackDefinition on removal}
}
\description{
Add / List / Remove ssffTrackDefinition to / from / of emuDB. 
An ssffTrack (often simply referred to as a track) references 
data that is stored in the Simple Signal File Format (SSFF) 
in the according bundle folders. The two most common types of data are:
\itemize{
\item{complementary data that was acquired during the recording 
such as data acquired during electromagnetic 
articulographic (EMA) or electropalatography (EPG) recordings;}
\item{derived data, i.e. data that was calculated from the original audio signal 
such as formant values and their bandwidths or the short-term Root Mean Square amplitude of the signal.}
}
For more information on the structural elements of an emuDB see \code{vignette(emuDB)}.
}
\seealso{
\link[emuR:AddListRemoveSsffTrackDefinition]{emuR::remove_ssffTrackDefinition}
}
