% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_emuR_re-export.R
\name{rename_emuDB}
\alias{rename_emuDB}
\title{Rename an emuDB}
\usage{
rename_emuDB(databaseDir, newName)
}
\arguments{
\item{databaseDir}{directory of the emuDB}

\item{newName}{new name of emuDB}
}
\description{
Rename a emuDB. This effectively renames the folder of a 
emuDB the _DBconfig.json file as well as the "name" entry in the _DBconfig.json
file and the _emuDBcache.sqlite file if available.
}
\seealso{
\link[emuR:rename_emuDB]{emuR::rename_emuDB}
}
