% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_sequence_ops_optimized.R
\name{retreat}
\alias{retreat}
\title{Retreat Backward in Sequence}
\usage{
retreat(.segments, steps_backward, ...)
}
\arguments{
\item{.segments}{segment_list or lazy_segment_list object}

\item{steps_backward}{Integer; number of steps to move backward (positive value)}

\item{...}{Additional arguments passed to scout()}
}
\value{
segment_list (or lazy_segment_list if collect=FALSE)
}
\description{
Convenience function for moving backward. Simply calls scout with negative offset.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Get previous phoneme
prev_phone <- retreat(segments, steps_backward = 1)

# Equivalent to:
prev_phone <- scout(segments, steps_forward = -1)
\dontshow{\}) # examplesIf}
}
