% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_transcription_system_optimized.R
\name{rewrite_annotations_optimized}
\alias{rewrite_annotations_optimized}
\title{Optimized annotation file rewriting}
\usage{
rewrite_annotations_optimized(corpus, session, bundle, verbose = TRUE)
}
\description{
Optimized annotation file rewriting
}
\keyword{internal}
