% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_transcription_system.R
\name{rewrite_annotations_parallel}
\alias{rewrite_annotations_parallel}
\title{Rewrite all annotations in parallel}
\usage{
rewrite_annotations_parallel(corpus, verbose = TRUE)
}
\description{
Rewrite all annotations in parallel
}
\keyword{internal}
