% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_transcription_system.R
\name{rewrite_bundle_annotation}
\alias{rewrite_bundle_annotation}
\title{Rewrite annotation file for a single bundle}
\usage{
rewrite_bundle_annotation(corpus, session, bundle, con, verbose = TRUE)
}
\description{
Rewrite annotation file for a single bundle
}
\keyword{internal}
