% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_signalextensions.R
\name{sampleRates}
\alias{sampleRates}
\title{Produce a table of sample rates of all media and SSFF track files}
\usage{
sampleRates(emuDBhandle)
}
\arguments{
\item{emuDBhandle}{An Emu database handle.}
}
\value{
a tibble with columns "session" and "bundle", and then one column for each SSFF track defined.
}
\description{
This function will access all defined SSFF track track files mentioned in an
SSFF track definition as well as the speech signal files, and extract and
list all sample rates. This function works from the actual track files and
not cached information, and may be slow on large databases with many SSFF
tracks defined, if stored on a media with slower file access speeds.
}
\examples{
reindeer:::create_ae_db(verbose = TRUE) -> emuDBhandle
samplerates(emuDBhandle)
reindeer:::unlink_emuRDemoDir()

}
\seealso{
reindeer::load_emuDB, emuR::load_emuDB
}
