% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_autosync.R
\name{save_sync_state}
\alias{save_sync_state}
\title{Save sync state}
\usage{
save_sync_state(db_handle, state)
}
\arguments{
\item{db_handle}{An emuDBhandle object}

\item{state}{List with sync state}
}
\description{
Save sync state
}
\keyword{internal}
