% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_core.R
\name{serialize_metadata_value}
\alias{serialize_metadata_value}
\title{Serialize a metadata value for storage}
\usage{
serialize_metadata_value(value)
}
\description{
Serialize a metadata value for storage
}
\keyword{internal}
