% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_emuR_re-export.R
\name{set_legalLabels}
\alias{set_legalLabels}
\title{Define which labels are legal for an attribute of a level. (A copy of the emuR function)}
\usage{
set_legalLabels(emuDBhandle, levelName, attributeDefinitionName, legalLabels)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link[emuR]{load_emuDB}}}

\item{levelName}{name of level}

\item{attributeDefinitionName}{name of attributeDefinition (can be and often is the level name)}

\item{legalLabels}{character vector of labels}
}
\description{
Set / Get / Remove legal labels of a specific attributeDefinition of a emuDB. 
The legal labels are a character vector of strings
that specifies the labels that are legal (i.e. allowed / valid) for the given attribute. 
As the EMU-webApp won't allow the annotator to enter any labels that are not 
specified in this array, this is a simple way of assuring that a level 
has a consistent label set. For more information 
on the structural elements of an emuDB see \code{vignette(emuDB)}.
Note that defining legal labels for an attributeDefinition does not imply that the 
existing labels are checked for being 'legal' in the emuDB.
}
\seealso{
\link[emuR:SetGetRemoveLegalLabels]{emuR::set_legalLabels}
}
