% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_emuR_re-export.R
\name{set_levelCanvasesOrder}
\alias{set_levelCanvasesOrder}
\title{Set the order of displayed levels for a perspective (from emuR)}
\usage{
set_levelCanvasesOrder(emuDBhandle, perspectiveName, order)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link[emuR]{load_emuDB}}}

\item{perspectiveName}{name of perspective}

\item{order}{character vector containing names of levelDefinitions}
}
\description{
Set / Get which levels of an emuDB to display as level canvases (in a 
given perspective of the EMU-webApp),
and in what order. Level canvases refer to levels of 
the type "SEGMENT" or "EVENT" that are displayed by the EMU-webApp. Levels 
of type "ITEM" can always be displayed using the hierarchy view of the
web application but can not be displayed as level canvases.
For more information on the structural elements of an emuDB 
see \code{vignette{emuDB}}.
}
\seealso{
\link[emuR:SetGetlevelCanvasesOrder]{emuR::set_levelCanvasesOrder}
}
