% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer-corpus.R
\name{set_metadata_bundle}
\alias{set_metadata_bundle}
\title{Set bundle-level metadata}
\usage{
set_metadata_bundle(corpus_obj, session_pattern, bundle_pattern, metadata_list)
}
\description{
Set bundle-level metadata
}
\keyword{internal}
