% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_metadata_io.R
\name{set_metadata_database}
\alias{set_metadata_database}
\title{Set database-level metadata}
\usage{
set_metadata_database(corpus_obj, metadata_list)
}
\description{
Set database-level metadata
}
\keyword{internal}
