% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_metadata_io.R
\name{set_metadata_session}
\alias{set_metadata_session}
\title{Set session-level metadata}
\usage{
set_metadata_session(corpus_obj, session_pattern, metadata_list)
}
\description{
Set session-level metadata
}
\keyword{internal}
