% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_database.R
\name{snapshot}
\alias{snapshot}
\title{Save the state of a speech database}
\usage{
snapshot(
  emuDBhandle,
  push.changes = TRUE,
  remote.name = "origin",
  remote.ref = "master"
)
}
\arguments{
\item{emuDBhandle}{The database handle}

\item{push.changes}{If \code{TRUE} all changes will pushed into the remote git server.}

\item{remote.name}{The name of the remote repository.}

\item{remote.ref}{The reference in the remote repository where changes should be pushed.}
}
\description{
This function saves a snapshot of an Emu speech database into a git repository and
optionally pushes changes to a remote git server. All altered files will be included in the snapshot.
}
\details{
If a remote server has been specified for the git repository, the changes will be pushed there into the "master" branch by default.
}
