% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_corpus_config.R
\name{store_DBconfig}
\alias{store_DBconfig}
\title{Store database configuration}
\usage{
store_DBconfig(obj, dbConfig, basePath = NULL)
}
\arguments{
\item{obj}{corpus object or emuDBhandle}

\item{dbConfig}{configuration list}

\item{basePath}{optional basePath override}
}
\description{
Store database configuration
}
\keyword{internal}
