% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_cache_system.R
\name{store_draft_annotations}
\alias{store_draft_annotations}
\title{Store draft annotations in cache}
\usage{
store_draft_annotations(
  con,
  session,
  bundle,
  level_name,
  level_type,
  attribute_name,
  annotations,
  parameters,
  error_occurred = FALSE,
  error_message = NULL
)
}
\arguments{
\item{con}{Database connection}

\item{session}{Session name}

\item{bundle}{Bundle name}

\item{level_name}{Level name}

\item{level_type}{Level type (SEGMENT, EVENT, or ITEM)}

\item{attribute_name}{Attribute name}

\item{annotations}{Data frame of annotations}

\item{parameters}{Parameters used for this bundle}

\item{error_occurred}{Did an error occur?}

\item{error_message}{Error message if error occurred}
}
\description{
Store draft annotations in cache
}
\keyword{internal}
