% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_transcription_system.R
\name{Suggestion}
\alias{Suggestion}
\title{Suggestion - Base class for transcription suggestions}
\usage{
Suggestion(
  corpus,
  session,
  bundle,
  level_name,
  level_type,
  attribute_name = level_name,
  suggestions,
  min_duration = 0.01,
  remove_empty = TRUE
)
}
\description{
Suggestion - Base class for transcription suggestions
}
