% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_lazy_segment_list.R
\name{summary.lazy_segment_list}
\alias{summary.lazy_segment_list}
\title{Summary method for lazy_segment_list}
\usage{
## S7 method for class <reindeer::lazy_segment_list>
summary(object, ...)
}
\arguments{
\item{object}{lazy_segment_list object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for lazy_segment_list
}
