% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_list_display.R
\name{summary.segment_list}
\alias{summary.segment_list}
\title{Summary method for segment_list}
\usage{
## S7 method for class <reindeer::segment_list>
summary(object, ...)
}
\arguments{
\item{object}{The segment_list object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for segment_list
}
