% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_autosync.R
\name{sync_annot_to_eaf}
\alias{sync_annot_to_eaf}
\title{Synchronize all _annot.json files to EAF format}
\usage{
sync_annot_to_eaf(
  db_handle,
  changed_bundles = NULL,
  align_items = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{db_handle}{An emuDBhandle object}

\item{changed_bundles}{Data frame with changed bundles, or NULL to check all}

\item{align_items}{Logical; whether to align ITEMs with time info}

\item{verbose}{Logical; show progress}
}
\description{
Synchronize all _annot.json files to EAF format
}
\keyword{internal}
