% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_autosync.R
\name{sync_database}
\alias{sync_database}
\title{Manually trigger synchronization}
\usage{
sync_database(
  db_handle,
  sync_eaf = TRUE,
  sync_cmdi = TRUE,
  force = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{db_handle}{An emuDBhandle object}

\item{sync_eaf}{Logical; sync EAF files}

\item{sync_cmdi}{Logical; sync CMDI file}

\item{force}{Logical; force sync even if no changes detected}

\item{verbose}{Logical; show progress}
}
\description{
Manually trigger synchronization
}
\keyword{internal}
