% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_transcription_system.R
\name{transcribe}
\alias{transcribe}
\title{Transcribe - Apply transcription suggestions to database}
\usage{
transcribe(suggestion, ...)
}
\arguments{
\item{suggestion}{A Suggestion object}

\item{force}{Skip confirmations and ignore errors}

\item{verbose}{Show progress messages}
}
\value{
TranscriptionLog object
}
\description{
Transcribe - Apply transcription suggestions to database
}
