% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_transcription_system_optimized.R
\name{transcribe_parallel}
\alias{transcribe_parallel}
\title{Transcribe multiple suggestion objects in parallel}
\usage{
transcribe_parallel(
  suggestion_list,
  n_cores = parallel::detectCores() - 1,
  verbose = TRUE
)
}
\arguments{
\item{suggestion_list}{List of Suggestion objects}

\item{n_cores}{Number of cores to use}

\item{verbose}{Show progress}
}
\description{
Transcribe multiple suggestion objects in parallel
}
