% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_transcription_system.R
\name{TranscriptionLog}
\alias{TranscriptionLog}
\title{TranscriptionLog - Tracks changes made by transcription}
\usage{
TranscriptionLog(corpus, session, bundle, level_name, operation = "transcribe")
}
\description{
Stores information about transcription operations to enable rollback
}
