% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_autosync.R
\name{with_auto_sync}
\alias{with_auto_sync}
\title{Wrapper for emuR functions that modify annotations}
\usage{
with_auto_sync(db_handle, fun, ...)
}
\arguments{
\item{db_handle}{An emuDBhandle object}

\item{fun}{Function to wrap (e.g., emuR::set_levelCanvasesOrder)}

\item{...}{Arguments to pass to the function}
}
\description{
This creates a wrapper that automatically syncs after annotation changes
}
\keyword{internal}
