% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_emuR_re-export.R
\name{write_bundleList}
\alias{write_bundleList}
\title{Store a list of bundles or segments for later use in the transcription (web) application}
\usage{
write_bundleList(
  emuDBhandle,
  name,
  bundleList,
  seglist,
  updateDBconfig = T,
  verbose = T
)
}
\arguments{
\item{emuDBhandle}{emuDB handle object (see \link[emuR]{load_emuDB})}

\item{name}{name of bundleList (excluding the _bundleList.json suffix)}

\item{bundleList}{tibble/data.frame with the columns \code{session}, \code{name}, 
\code{comment} (optional), \code{finishedEditing} (optional). Use \link[emuR]{list_bundles}}

\item{seglist}{segment list returned by \link[emuR]{query} function. If set the 
\code{bundleList} parameter will be ignored and a bundleList will be created by
collapsing the segments as timeAnchors into the \code{_bundleList.json}}

\item{updateDBconfig}{if set to TRUE (the default) DBconfig will be updated 
with the fields}

\item{verbose}{be verbose
\itemize{
 \item \code{"bundleComments": true}
 \item \code{"bundleFinishedEditing": true}
}}
}
\description{
write bundleList JSON file to emuDB
}
\details{
Write bundleList JSON file to emuDB sub-dir \code{bundleLists/}
}
\seealso{
\link[emuR:write_bundleList]{emuR::write_bundleList}
}
