% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_metadata_optimized.R
\name{write_metadata_to_json}
\alias{write_metadata_to_json}
\title{Write metadata to .meta_json file (ground truth)}
\usage{
write_metadata_to_json(corpus_obj, meta_list, session, bundle, level)
}
\description{
Write metadata to .meta_json file (ground truth)
}
\keyword{internal}
