% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_metadata_optimized.R
\name{add_digests}
\alias{add_digests}
\title{Add file digests to metadata}
\usage{
add_digests(
  corpus_obj,
  sessionPattern = ".*",
  bundlePattern = ".*",
  algorithm = "sha1"
)
}
\arguments{
\item{corpus_obj}{Corpus object}

\item{sessionPattern}{Session pattern}

\item{bundlePattern}{Bundle pattern}

\item{algorithm}{Hash algorithm}
}
\description{
Add file digests to metadata
}
