% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{add_digests}
\alias{add_digests}
\title{Add identifying information based on the content of the wave file to the metadata information for the bundle.}
\usage{
add_digests(dbhandle, sessionPattern = ".*", bundlePattern = ".*",
  algorithm = "sha1")
}
\arguments{
\item{dbhandle}{The handle for the emuR database.}

\item{sessionPattern}{A regexp pattern that allows you to limit which sessions should be affected by the manibpulation.}

\item{bundlePattern}{A regexp pattern that allows you to limit which bundles to include.}

\item{algorithm}{The name of the hashing algorithm, according to the \code{\link[digest]{digest}} function.}
}
\description{
This function will extract information from the wav file associated with a bundle, and add it to the set of metadata
for the bundle. This information can later be used to verify that the file has not been altered later on, or to deidentify
wav files in a reversable manner for use outside of the emuR framework. Deidentified files are sometimes useful for blinded randomized
perceptual testing, and the ability to reverse the procedure is then essential to link the results of the evaluation back to the original
recording extracted from the emuR data base.
}
\examples{
\dontrun{
create_ae_db() -> ae
add_digests(ae)
get_metadata(ae,Excelfile = NULL) -> res
print(res)
unlink_emuRDemoDir()
}

}
