% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_metadata.R,
%   R/reindeeR_metadata_optimized.R
\name{add_metadata}
\alias{add_metadata}
\title{A utility function used for programatically setting metadata for a bundle, or
default values for a session or an entire database.}
\usage{
add_metadata(
  corpus_obj,
  metadataList,
  session = NULL,
  bundle = NULL,
  reset.before.add = FALSE
)

add_metadata(
  corpus_obj,
  metadataList,
  session = NULL,
  bundle = NULL,
  reset.before.add = FALSE
)
}
\arguments{
\item{corpus_obj}{A corpus object}

\item{metadataList}{Named list of metadata}

\item{session}{Optional session name}

\item{bundle}{Optional bundle name}

\item{reset.before.add}{Clear existing metadata first}

\item{emuDBhandle}{An Emu database handle}
}
\description{
The function takes a list and a specification of where the metadata should be
set. The default behaviour is to keep already set metadata, and overwrite
only the values that are named in the list. The user may change this
behaviour by setting \code{reset.before.add=TRUE}, in which case all previous
bundle, session or database level metadata will be replaced with the contents
of the list.
}
\details{
If a bundle name and a \code{session} name is provided, the metadata will be
inserted only for that fully specified \code{bundle}. If only a \code{bundle}
name is provided, the function will add the metadata for the bundle only if
there is just one session in the database. If there are multiple
\code{session}s, the function will given an error.

If no \code{session} or \code{bundle} names are provided, the metadata will
be inserted as default values for the entire database. Please note that
database wide ingested metadata defaults currently interferes with the Emu
web app and should not be put into general use before this issue has been
fixed in the Emu web app.
}
\examples{
\dontrun{
emuR::create_emuRdemoData()
ae_test <- load_emuDB(file.path(tempdir(),"emuR_demoData","ae_emuDB"))

# Database-wide default information
add_metadata(ae_test,list("Accent"="Northern","Elicitation"="Scripted"))
#Bundle specific information
add_metadata(ae_test,list("Speaker.Sex"="Male","Date"="2020-03-04"),session="0000",bundle="msajc003")
get_metadata(ae_test) -> res
print(res)
}

}
