% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_metadata_optimized.R
\name{add_metadata}
\alias{add_metadata}
\title{Add metadata programmatically}
\usage{
add_metadata(
  corpus_obj,
  metadataList,
  session = NULL,
  bundle = NULL,
  reset.before.add = FALSE
)
}
\arguments{
\item{corpus_obj}{A corpus object}

\item{metadataList}{Named list of metadata}

\item{session}{Optional session name}

\item{bundle}{Optional bundle name}

\item{reset.before.add}{Clear existing metadata first}
}
\description{
Add metadata programmatically
}
