% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{add_metadata}
\alias{add_metadata}
\title{A utility function used for setting metadata for a bundle, at for a session or for an entire database.}
\usage{
add_metadata(
  dbhandle,
  metadataList,
  bundle = NULL,
  session = NULL,
  reset.before.add = FALSE
)
}
\arguments{
\item{dbhandle}{An Emu database handle}

\item{metadataList}{A list specifying the metadata to be set. If set to an empty list (\code{list()}) the function will clear all metadata, if the argument \code{reset.before.add=TRUE} is given by the user.}

\item{bundle}{An optional name of a bundle}

\item{session}{An optional name of a session}

\item{reset.before.add}{If set to TRUE, the function will ignore previously set metadata and simply add the metadata supplied in the list.}
}
\value{

}
\description{
The function takes a list and a specification of where the metadata should be set. The default behaviour is to
keep already set metadata, and overwrite only the values that are names  in the list. The user may change this
behaviour by setting \code{reset.before.add=TRUE}, in which case all previous metadata will be replaced with the
data given in the list.
}
\details{
If bundle name and a \code{session} name is provided, the metadata will be inserted only for that fully speficied \code{bundle}.
If only a \code{bundle} name is provided, the function will add the metadata for the bundle only if there is just one session
in the database. If there are multiple \code{session}s, the function will given an error.

If no \code{session} or \code{bundle} names are provided, the metadata will be inserted as default values for the entire database.
}
