% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_annotate.R
\name{annotate_periods}
\alias{annotate_periods}
\title{Mark periods as events}
\usage{
annotate_periods(
  emuDBhandle,
  beginTime = 0,
  endTime = 0,
  windowShift = 5,
  minF = 75,
  maxF = 600,
  windowShape = "Gaussian1",
  interpolation = "cubic",
  RelativeWidth = 1,
  levelname = "AUTO_PERIODS",
  clear = TRUE
)
}
\arguments{
\item{emuDBhandle}{A emuR database handle.}

\item{beginTime}{The start time of the section of the sound file that should be processed.}

\item{endTime}{The end time of the section of the sound file that should be processed.}

\item{windowShift}{The measurement interval (frame duration), in seconds. If you supply 0, Praat will use a time step of 0.75 / (pitch floor), e.g. 0.01 seconds if the pitch floor is 75 Hz; in this example, Praat computes 100 pitch values per second.}

\item{minF}{candidates below this frequency will not be recruited. This parameter determines the effective length of the analysis window: it will be 3 longest periods long, i.e., if the pitch floor is 75 Hz, the window will be effectively 3/75 = 0.04 seconds long. Note that if you set the time step to zero, the analysis windows for consecutive measurements will overlap appreciably: Praat will always compute 4 pitch values within one window length, i.e., the degree of oversampling is 4.}

\item{maxF}{Candidates above this frequency will be ignored.}

\item{windowShape}{the analysis window function used when extracting part of a sound file for analysis. De faults to "Hanning".}

\item{interpolation}{The method used for finding intensity associated with the pitch period event (ont of "nearest", or "linear", "cubic", "sinc70" or "sinc700" interpolation).}

\item{RelativeWidth}{The relative with of the window used for extracting part of the signal for analysis.}

\item{levelname}{The name of the (EVENT) annotation level to insert into the database.}

\item{clear}{Should the \code{levename} level be cleared before inserting annotations?}
}
\value{
A \link[tibble:tibble]{tibble::tibble} containing information on the labels inserted (session, bundle, start, level, attribute, labels)
}
\description{
This function uses Praat \insertCite{Praatdoingphoneti:2022}{reindeer} to
extract a periodicity contour, and the associated intensity. The estimated
timing of each glottal cycle is marked at the maximum peak, and the timing
and intensity of that peak is then marked as an event with associated label in the
\code{levelname} transcription level.
}
\details{
If an intensity level could not be determined at the peak, the label of the
EVENT will be empty.
}
\references{
\insertAllCited{}
}
