% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_lazy_segment_list.R
\name{apply_ascend_transform}
\alias{apply_ascend_transform}
\title{Apply Ascend (Dominance Upward) Transform}
\usage{
apply_ascend_transform(sql, level)
}
\description{
Apply Ascend (Dominance Upward) Transform
}
\keyword{internal}
