% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_lazy_segment_list.R
\name{apply_descend_transform}
\alias{apply_descend_transform}
\title{Apply Descend (Dominance Downward) Transform}
\usage{
apply_descend_transform(sql, level)
}
\description{
Apply Descend (Dominance Downward) Transform
}
\keyword{internal}
