% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_lazy_segment_list.R
\name{apply_retreat_transform}
\alias{apply_retreat_transform}
\title{Apply Retreat (Backward Sequence) Transform}
\usage{
apply_retreat_transform(query, n = 1)
}
\arguments{
\item{query}{list(sql, params) -- current query}

\item{n}{integer -- number of steps backward}
}
\value{
list(sql, params) -- modified query
}
\description{
Apply Retreat (Backward Sequence) Transform
}
\keyword{internal}
