% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_lazy_segment_list.R
\name{apply_retreat_transform}
\alias{apply_retreat_transform}
\title{Apply Retreat (Backward Sequence) Transform}
\usage{
apply_retreat_transform(sql, n = 1)
}
\description{
Apply Retreat (Backward Sequence) Transform
}
\keyword{internal}
