% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_lazy_segment_list.R
\name{apply_scout_transform}
\alias{apply_scout_transform}
\title{Apply Scout (Forward Sequence) Transform}
\usage{
apply_scout_transform(sql, n = 1)
}
\description{
Apply Scout (Forward Sequence) Transform
}
\keyword{internal}
