% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_lazy_segment_list.R
\name{apply_transform}
\alias{apply_transform}
\title{Apply a Transform to SQL Query}
\usage{
apply_transform(sql, transform, verbose = FALSE)
}
\arguments{
\item{sql}{Current SQL query}

\item{transform}{List with type and parameters}

\item{verbose}{Logical}
}
\value{
Modified SQL query
}
\description{
Apply a Transform to SQL Query
}
\keyword{internal}
