% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_segment_list.R
\name{as_segment_list}
\alias{as_segment_list}
\title{Convert data.frame to segment_list}
\usage{
as_segment_list(x, db_uuid = NULL, db_path = NULL)
}
\description{
Convert data.frame to segment_list
}
