% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_core.R
\name{assess}
\alias{assess}
\title{Initialize track simulation cache database}
\usage{
assess(
  segment_list,
  simulation_results,
  .metric = yardstick::rmse,
  .track_name = NULL,
  .detailed = FALSE,
  .verbose = TRUE
)
}
\description{
Initialize track simulation cache database
}
\keyword{internal}
