% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_autosync.R
\name{auto_sync_check}
\alias{auto_sync_check}
\title{Check and sync after database operations}
\usage{
auto_sync_check(db_handle)
}
\arguments{
\item{db_handle}{An emuDBhandle object}
}
\description{
This function should be called after operations that modify the database
}
\keyword{internal}
