% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_metadata.R,
%   R/reindeeR_metadata_optimized.R
\name{biographize}
\alias{biographize}
\title{Create a biography of the labels in a list of segments in a tidy manner}
\usage{
biographize(segs_tbl, corpus_obj, compute_digests = FALSE, algorithm = "sha1")

biographize(segs_tbl, corpus_obj, compute_digests = FALSE, algorithm = "sha1")
}
\arguments{
\item{segs_tbl}{Tibble from query}

\item{corpus_obj}{Corpus object}

\item{compute_digests}{Compute file checksums}

\item{algorithm}{Hash algorithm}

\item{emuDBhandle}{A \code{\link{emuR}} database handle.}
}
\value{
A \code{\link[dplyr]{tibble}}
}
\description{
Create a biography of the labels in a list of segments in a tidy manner

Add biographical metadata to segment list (biographize)
}
\examples{
\dontrun{
create_emuRdemoData()
ae_test <- load_emuDB(file.path(tempdir(),"emuR_demoData","ae_emuDB"))

# Database-wide default information
add_metadata(ae_test,list("Accent"="Northern","Elicitation"="Scripted"))
#Bundle specific information
add_metadata(ae_test,list("Speaker.Sex"="Male","Date"="2020-03-04"),session="0000",bundle="msajc003")

# Get all the 'n' segments in the database
query(ae_test,"Phonetic = n",resultType = "tibble") -> ae_nt
# Add information related to the nature the recording sessions
# e.g. the speaker ID, the date of the recording
ae_nt \%>\% biographize(ae_test) \%>\% glimpse()
# This code does the same as the above, but it will also compute new
# information that is strictly  aimed at identifying the recording
# (length of recording (in ms) and a sha1 digest of the wav file).
ae_nt \%>\%
   biographize(ae_test,compute_digests=TRUE,algorithm="sha1") \%>\%
   glimpse()
rm(ae_test)

}

}
