% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_import_export.R
\name{biographize}
\alias{biographize}
\title{Add biographical metadata to segment list (biographize)}
\usage{
biographize(segs_tbl, corpus_obj, compute_digests = FALSE, algorithm = "sha1")
}
\arguments{
\item{segs_tbl}{Tibble from query}

\item{corpus_obj}{Corpus object}

\item{compute_digests}{Compute file checksums}

\item{algorithm}{Hash algorithm}
}
\description{
Add biographical metadata to segment list (biographize)
}
