% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_metadata.R
\name{biographize}
\alias{biographize}
\title{Create a biography of the labels in a list of segments in a tidy manner}
\usage{
biographize(segs_tbl, emuDBhandle, compute_digests = FALSE, algorithm = "sha1")
}
\arguments{
\item{segs_tbl}{The \code{\link[dplyr]{tibble}} that is the result \code{\link[emuR]{query}} call.}

\item{emuDBhandle}{A \code{\link{emuR}} database handle.}

\item{compute_digests}{Should information that describes the recorded sound files be computed so that is is definitelly part of the
added metadata information.}

\item{algorithm}{The checksum algorithm that should be used when computing sound file information.}
}
\value{
A \code{\link[dplyr]{tibble}}
}
\description{
Create a biography of the labels in a list of segments in a tidy manner
}
\examples{
\dontrun{
create_emuRdemoData()
ae <- load_emuDB(file.path(tempdir(),"emuR_demoData","ae_emuDB"))

# Database-wide default information
add_metadata(ae,list("Accent"="Northern","Elicitation"="Scripted"))
#Bundle specific information
add_metadata(ae,list("Speaker.Sex"="Male","Date"="2020-03-04"),session="0000",bundle="msajc003")

# Get all the 'n' segments in the database
query(ae_test,"Phonetic = n",resultType = "tibble") -> ae_nt
# Add information related to the nature the recording sessions
# e.g. the speaker ID, the date of the recording
ae_nt \%>\% biographize(ae_test) \%>\% glimpse()
# This code does the same as the above, but it will also compute new
# information that is strictly  aimed at identifying the recording
# (length of recording (in ms) and a sha1 digest of the wav file).
ae_nt \%>\%
   biographize(ae_test,compute_digests=TRUE,algorithm="sha1") \%>\%
   glimpse()
rm(ae)
unlink(file.path(tempdir(),"emuR_demoData"),recursive=TRUE)
}

}
