% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_metadata.R
\name{biographize}
\alias{biographize}
\title{Create a biography of the labels in a list of segments in a tidy manner}
\usage{
biographize(segs_tbl, emuDBhandle, compute_digests = FALSE, algorithm = "sha1")
}
\arguments{
\item{segs_tbl}{The \code{\link[dplyr]{tibble}} that is the result \code{\link[emuR]{query}} call.}

\item{emuDBhandle}{A \code{\link{emuR}} database handle.}

\item{compute_digests}{Should information that describes the recorded sound files be computed so that is is definitelly part of the
added metadata information.}

\item{algorithm}{The checksum algorithm that should be used when computing sound file information.}
}
\value{
A \code{\link[dplyr]{tibble}}
}
\description{
Create a biography of the labels in a list of segments in a tidy manner
}
\examples{
\dontrun{
## This code just sets up a new emuR database and inserts some
## fake metadata into it.
create_ae_db() -> ae_test
make_dummy_metafiles(ae_test)
# Get all the 'n' segments in the database
query(ae_test,"Phonetic = n",resultType = "tibble") -> ae_nt
# Add information related to the nature the recording sessions
# e.g. the speaker ID, the date of the recording
ae_nt \%>\% biographize(ae_test)
# This code does the same as the above, but it will also compute new
# information that is strictly  aimed at identifying the recording
# (length of recording (in ms) and a sha1 digest of the wav file).
ae_nt \%>\% biographize(ae_test,compute_digests=TRUE,algorithm="sha1")
}

}
