% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_executor.R, R/reindeer_query_optimized.r
\name{build_base_sql}
\alias{build_base_sql}
\title{Build Base SQL Query Without Execution}
\usage{
build_base_sql(db_path, parsed, opts = list())

build_base_sql(db_path, parsed, opts = list())
}
\arguments{
\item{db_path}{Path to SQLite database}

\item{parsed}{Parsed EQL query}

\item{opts}{Additional options}
}
\value{
SQL query string

SQL query string
}
\description{
Extracts the SQL query that would be executed, for use in lazy evaluation.

Extracts the SQL query that would be executed, for use in lazy evaluation.
}
\keyword{internal}
