% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer-corpus.R
\name{build_emuDB_cache}
\alias{build_emuDB_cache}
\title{Corpus cache file update}
\usage{
build_emuDB_cache(
  database_dir,
  parallel = TRUE,
  workers = future::availableCores() - 1,
  batch_size = 50,
  verbose = TRUE
)
}
\arguments{
\item{database_dir}{Path to the Emu-DBMS corpus base directory}

\item{parallel}{Use parallel processing (default: TRUE)}

\item{workers}{Number of parallel workers (default: all but one available cores are used)}

\item{batch_size}{Number of bundles to process in each batch for DB insertion}

\item{verbose}{Show progress information}
}
\value{
A corpus object
}
\description{
Builds an Emu-DBMS compliant cache file for a corpus.
}
