% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_lazy_segment_list.R
\name{build_sql_from_parts}
\alias{build_sql_from_parts}
\title{Build SQL Query from Query Parts}
\usage{
build_sql_from_parts(query_parts, verbose = FALSE)
}
\arguments{
\item{query_parts}{List of query components}

\item{verbose}{Logical}
}
\value{
SQL query string
}
\description{
Build SQL Query from Query Parts
}
\keyword{internal}
