% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_class.R
\name{bundle_list}
\alias{bundle_list}
\title{Bundle List Class - Tibble with session/bundle information and metadata}
\usage{
bundle_list(
  .data = data.frame(session = character(), bundle = character(), stringsAsFactors =
    FALSE)
)
}
\description{
An S7 class that extends tibble to represent a list of bundles with
their associated metadata following inheritance rules.
}
