% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_database.R, R/metadata_core.R
\name{deserialize_metadata_value}
\alias{deserialize_metadata_value}
\title{Deserialize metadata value}
\usage{
deserialize_metadata_value(value_str, type_str)

deserialize_metadata_value(value_str, type_str)
}
\description{
Deserialize metadata value

Deserialize a metadata value from storage
}
\keyword{internal}
