% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_methods.R
\name{get_handle}
\alias{get_handle}
\title{Get or restore database handle for corpus}
\usage{
get_handle(corpus_obj, verbose = FALSE)
}
\arguments{
\item{corpus_obj}{A corpus object}

\item{verbose}{Whether to show loading messages}
}
\value{
An emuDBhandle with valid connection
}
\description{
Get or restore database handle for corpus
}
