% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{get_metadata}
\alias{get_metadata}
\title{This function gathers metata for bundles in an emuR database.}
\usage{
get_metadata(dbhandle, Excelfile = NULL,
  add.metadata = c("Session.DateTime", "Speaker.ID"),
  overwrite = FALSE, session = ".*")
}
\arguments{
\item{dbhandle}{The database handle of an emuR database.}

\item{Excelfile}{The full path and file name of the Excel file that the metadata should be written to. The function will not overwrite this file, unless \code{overwrite} is set to \code{TRUE}.}

\item{add.metadata}{A vector of column names that the function should make sure
that they are present in the output.}

\item{overwrite}{The default behaviour is that an Excel file should not be
overwritten if it is present already. If this parameter is \code{TRUE} then the file will be overwritten.}

\item{session}{A session pattern. Used for editing only the metadata of bundles in a
specific  session.}
}
\value{
A data frame containing inforamtion about the 'meta_json' files found
\describe{
  \item{session}{The name of the session.}
  \item{bundle}{The bundle name}
  \item{file}{The file name of the meta_json file found in the database}
  \item{absolute_file_path}{The full absolute path to the 'meta_json' file.}
}
In addition, the \code{\link[base]{data.frame}} will contain one column for every
type of information given in any of the 'meta_json' files.
}
\description{
Metadata of a file is stored in 'meta_json' files. This function goes through all
bundles, parses the JSON data and collects everything into a data fram which is
returned. The structure of the metadata does not have to be consistent across
meta_json files. New columns are added to the data.frame as new fields are detected.
And the used may specify addidional columns that should be added to the data frame
regardless of them being present in any fo the 'meta_json' files.
}
\details{
The user may also give the name of an Excel file to which the metadata table should be
exported. The session and bundle columns of this file will be locked to make it
easier to edit the metadata. The two file columns are also hidden.
}
\examples{
\\donotrun{
create_ae_db() -> ae
make_dummy_metafiles(ae)
get_metadata(ae)
## Some cleanup code
unlink_emuRDemoDir()
rm(ae)
}

}
