% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_signalextensions.R
\name{get_metaFuncFormals}
\alias{get_metaFuncFormals}
\title{A utility function that builds a list of parameters to use in function call based on metadata}
\usage{
get_metaFuncFormals(
  emuDBhandle,
  session,
  bundle,
  onTheFlyFunctionName,
  onTheFlyParams = list(),
  metadata.defaults = list(Gender = NA, Age = 35),
  package = "superassp"
)
}
\arguments{
\item{emuDBhandle}{The database handle}

\item{session}{Only consider this session.}

\item{bundle}{The bundle in the session to build an argument list for.}

\item{onTheFlyFunctionName}{The name of the function for which appropriate
parameters should be deduced.}

\item{onTheFlyParams}{A list of explicit DSP function parameters.}

\item{metadata.defaults}{Default settings for metadata. This argument is
primarily used for setting default Age and Gender for bundles with no such
metadata already set.}

\item{package}{The signal processing package where the \code{onTheFlyFunctionName}
function may be found. Defaults to \link{superassp}.}
}
\value{
A list which may be supplied to a \code{do.call} call.
}
\description{
This function takes a function and explicit argument list and builds a list of parameters for a function call considering the metadata of a bundle (in a session).
}
\seealso{
metadata_parameters

do.call
}
