% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_signalextensions.R
\name{get_parameters}
\alias{get_parameters}
\title{Provides the user with speaker dependent signal processing parameters}
\usage{
get_parameters(recompute = FALSE)
}
\arguments{
\item{recompute}{If \code{TRUE}, this function will estimate age appropriate settings from the collection of empirical defaults stored in this package. If \code{FALSE}, this function will just return the stored \code{DSPP} data.}
}
\value{
A \link{tibble:tibble} with one row per age and gender combination, and with one column per parameter for which an age dependent setting was identified in the set of parameters.
}
\description{
The source of the default signal processing parameters may be a spreadsheet
file in the OOXML Workbook ISO/IEC 29500:2008 standard format.The simplest
way to obtain such a file is to provide this function with a file name that
does not exist. This function will then write the \link{DSPP} set to the file, and
the user may then edit the file by hand using Microsoft Excel, Libreoffice
Calc or some other standard compliant software. Once edited, this function
may be used to read in the spreadsheet file to a tibble that may be used for
signal processing purposes.
}
\details{
Alternatively, the user may use this function to just get the defaults stored
in \link{DSPP} returned directly by not supplying a file name.
}
