% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_autosync.R, R/reindeeR_cmdi.R,
%   R/reindeeR_metadata_optimized.R, R/reindeer-corpus.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
x \%||\% y

x \%||\% y

x \%||\% y

x \%||\% y
}
\description{
Null-coalescing operator

Null coalescing operator

Null coalescing operator

Null coalescing operator
}
\keyword{internal}
