% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_database.R, R/metadata_import_export.R,
%   R/reindeer_autosync.R, R/reindeer_cmdi.R
\name{\%||\%}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
a \%||\% b

a \%||\% b

a \%||\% b

a \%||\% b
}
\description{
Null coalescing operator

Null coalescing operator

Null-coalescing operator

Null coalescing operator
}
\keyword{internal}
