% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_import_export.R
\name{import_metadata}
\alias{import_metadata}
\title{Import metadata from Excel file}
\usage{
import_metadata(corpus_obj, Excelfile)
}
\arguments{
\item{corpus_obj}{A corpus object}

\item{Excelfile}{Path to Excel file to import}
}
\description{
Import metadata from Excel file
}
