% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{import_metadata}
\alias{import_metadata}
\title{Functions to import or add metadata information to database bundles.}
\usage{
import_metadata(dbhandle, Excelfile)
}
\arguments{
\item{dbhandle}{The emuR database handle of the database.}

\item{Excelfile}{The path to a properly formated Excel (.xlsx) file.}
}
\value{
A vector of .meta_json files updated by the call. The path for each file is given relative to the base of the EmuR database.
}
\description{
The function \code{import_metadata} takes an appropriately structured Excel file and uses the
gathered information to set metadata for bundles.
}
\details{
The first sheet ("bundles") in the Excel file should begin with the folowing two columns:
\itemize{
\item session
\item bundle
}
and then go on to have some columns which contains the metadata. Each row in the
data contains the information and metadata for a bundle (in a specific session).
The simples way to get an appropriately structed Excel file is to create one from a database using the
\code{\link{export_metadata}} function on an existing database and given an output file.

Please be aware that bundles that are speficied in the Excel file will have
their metadata files (ending with '.meta_json') overwritten when using the
\code{import_metadata}. So, please make sure to remove rows of bundles that should
not be altered from the Excel file before importing the metadata from it.
}
