% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{import_metadata}
\alias{import_metadata}
\title{A function for importing metadata from an Excel file}
\usage{
import_metadata(dbhandle, Excelfile, ignore.columns = NULL)
}
\arguments{
\item{dbhandle}{The emuR database handle of the database.}

\item{Excelfile}{The path to a properly formated Excel (.xlsx) file.}

\item{ignore.columns}{Columns in the Excel file (other than \code{session},\code{bundle},
\code{file} and \code{absolute_file_path}) that should not be considered to contain metadata.}
}
\value{
A vector of 'meta_json' files updated by the call. The path for each file is given relative to the base of the emuR database.
}
\description{
This function takes an Excel file (in .xlsx format) and extracts metadata from it.
The file should begin with the folowing columns:
\itemize{
\item session
\item bundle
\item file
\item absolute_file_path
}
and then go on to have some columns which contains the meta data. Each row in the
data contains the information and metadata for a bundle (in a specific session).
The simples way to get such a file is to create one from a database using the
\code{\link{get_metadata}} function. The fields \code{file} and
\code{absolute_file_path} will be hidden in this file to fascilitate editing of the
metadata.
}
\details{
Please be aware that bundles that are speficied in the Excel file will have
their metadata files (ending with '.meta_json') overwritten when using the
\code{import_metadata}. So, please make sure to remove rows of bundles that should
not be altered from the Excel file before importing the metadata from it.
}
