% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_database.R
\name{import_recordings}
\alias{import_recordings}
\title{Import a speech recordings into an Emu database session}
\usage{
import_recordings(
  emuDBhandle,
  dir,
  targetSessionName = "0000",
  speech.channel = 1,
  egg.channel = NULL,
  verbose = TRUE
)
}
\arguments{
\item{emuDBhandle}{The Emu database handle.}

\item{dir}{The directory containing speech recordings or session directories.}

\item{targetSessionName}{The default session name, if not specified using sub-directories in the import folder.}

\item{speech.channel}{The channel number containing the audio track to be stored in the database.}

\item{egg.channel}{An optional electroglottography channel number, which will
be stored in a separate ".egg" file.}

\item{verbose}{Should additional information and progress bar be displayed to the user?}
}
\description{
This function imports speech recordings stored in a directory into session in
an Emu database. The new session will be named according to the name the user
provides using the \code{targetSessionName} argument. A new bundle will be created
for every file with the file extension specified in the "mediafileExtension"
setting in the database configuration file (usually "wav"). ALternatively,
the user may set up sub-directories in the import folder, which will then be
used as recording session names.
}
\details{
In order to ensure that the database always contains a mono sound file, the
single channel indicated by the \code{speech.channel} argument will be extracted
and stored in the database.

The function also makes sure that 'OSCI' and 'SPEC' perspectives are defined
for the database.

The user may also indicate that an additional channel of the speech
recording, containing an electroglottography (EGG) track, should be stored in
an '.egg' file by indicating its channel number in the \code{egg.channel}
argument. The EGG track is often stored in the second channel of the speech
recording file, which means that \code{egg.channel=2} is usually appropriate. The
user should however verify this and a sound editor before importing speech
recordings.
}
