% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_simulation.R
\name{initialize_track_simulation_cache}
\alias{initialize_track_simulation_cache}
\title{Initialize track simulation cache database}
\usage{
initialize_track_simulation_cache(cache_dir, timestamp, dsp_function_name)
}
\description{
Initialize track simulation cache database
}
\keyword{internal}
