% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_segment_list.R
\name{is_extended_segment_list}
\alias{is_extended_segment_list}
\title{Check if object is an extended_segment_list}
\usage{
is_extended_segment_list(x)
}
\description{
Check if object is an extended_segment_list
}
