% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_lazy_segment_list.R
\name{lazy_segment_list}
\alias{lazy_segment_list}
\title{S7 class for lazy segment lists}
\usage{
lazy_segment_list(
  corpus = NULL,
  query_parts = list(),
  db_path = "",
  db_uuid = "",
  materialized = FALSE,
  cache = NULL
)
}
\description{
S7 class for lazy segment lists
}
