% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_emuR_re-export.R
\name{list_sessions}
\alias{list_sessions}
\title{Re-exported function from emuR list sessions in an emuDB. (A copy of the emuR function)}
\usage{
list_sessions(emuDBhandle, sessionPattern = ".*")
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link[emuR]{load_emuDB}}}

\item{sessionPattern}{A regular expression pattern matching session names to 
be searched for in the database. Note: "_ses$" is appended to this RegEx automatically}
}
\value{
data.frame object with session names
}
\description{
List session names of emuDB
}
\seealso{
\link[emuR:list_sessions]{emuR::list_sessions}
}
