% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_simulation.R
\name{list_simulations}
\alias{list_simulations}
\title{List available simulations in cache directory}
\usage{
list_simulations(cache_dir)
}
\arguments{
\item{cache_dir}{Directory containing simulation caches}
}
\value{
data.table with simulation metadata
}
\description{
List available simulations in cache directory
}
