% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR_transcriptionWork.R
\name{list_untranscribedBundles}
\alias{list_untranscribedBundles}
\title{Find bundles in a database that have not been transcribed yet.}
\usage{
list_untranscribedBundles(emuDBhandle)
}
\arguments{
\item{emuDBhandle}{The database handle of an \link{emuR} database.}
}
\value{
A vector of bundle names
}
\description{
Find bundles in a database that have not been transcribed yet.
}
\examples{
\dontrun{
reindeer:::create_ae_db() -> ae
# Make a test case for this particular function
unlink(file.path(ae$basePath,"0000_ses/msajc012_bndl/msajc012_annot.json"))
print(list_untranscribedBundles(ae))
reindeer:::detach_demo_db(ae)
}

}
\author{
Fredrik Karlsson
}
