% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_corpus_config.R
\name{load_DBconfig}
\alias{load_DBconfig}
\title{Load database configuration}
\usage{
load_DBconfig(obj)
}
\arguments{
\item{obj}{Either a character (basePath) or corpus object}
}
\value{
List with database configuration
}
\description{
Load database configuration
}
