% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_signalextensions.R
\name{match_parameters}
\alias{match_parameters}
\title{Metadata-aware selection of DSP parameters}
\usage{
match_parameters(
  emuDBhandle,
  onTheFlyFunctionName,
  metadata.defaults = list(Gender = NA, Age = 35),
  wide.format = FALSE,
  package = "superassp"
)
}
\arguments{
\item{emuDBhandle}{The database handle}

\item{onTheFlyFunctionName}{The name of the function for which appropriate parameters should be deduced.}

\item{metadata.defaults}{Default settings for metadata. This argument is primarily used for setting default Age and Gender for bundles with no such metadata already set.}

\item{wide.format}{The default output format of this function is one row per parameter and value pair. If \code{TRUE}, this argument forces the function to instead return one row per bundle in the database, and with parameter values in separate columns instead.}

\item{package}{The signal processing package where the \code{onTheFlyFunctionName} function may be found. Defaults to \link{superassp}.}
}
\value{
A data frame with columns \code{session}, \code{bundle}, \code{Parameter}, and \code{Setting}, grouped by \code{session} and \code{bundle}.
}
\description{
This function deduces the appropriate parameters
for a speech signal processing function based on the
speaker Age and Gender metadata set for a bundle.
}
\examples{
\dontrun{
reindeer:::unlink_emuRDemoDir()
reindeer:::create_ae_db() -> emuDBhandle
reindeer:::make_dummy_metafiles(emuDBhandle)
print(match_parameters(emuDBhandle,onTheFlyFunctionName = "forest") )
}
}
