% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_signalextensions.R
\name{match_parameters}
\alias{match_parameters}
\title{Select superassp DSP function parameters based on metadata}
\usage{
match_parameters(
  emuDBhandle,
  onTheFlyFunctionName,
  onTheFlyParams = NULL,
  metadata.defaults = list(Gender = NA, Age = 35),
  recompute = FALSE,
  package = "superassp"
)
}
\arguments{
\item{emuDBhandle}{The Emu database handle.}

\item{onTheFlyFunctionName}{The name of the digital speech processing function}

\item{onTheFlyParams}{If default parameters are given, they are applied to all bundles / files, overwriting parameters set based on metadata.}

\item{metadata.defaults}{Default values for metadata. Usually used for setting Age and Gender for bundles missing that information.}

\item{recompute}{If \code{TRUE} the Age and Gender specific DSP settings will be recomputed from the database. Please see \link{get_metadataParameters} for more information.}

\item{package}{}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with one row per bundle / session and with extra columns containing parameters to apply.
}
\description{
This function takes an Emu database handle and a function defined in
\link{superassp} or a named package, and outputs a \link[tibble:tibble]{tibble::tibble} containing
function paramers to use for each bundle. The function parameters are derived
from the database of age and gender specific settings obtained from in the
literature based Gender and Age metadata information set for the bundle.
}
\examples{
retuire(superassp)
reindeer:::create_ae_db(verbose = TRUE) -> emuDBhandle
reindeer:::add_dummy_metadata(emuDBhandle)
 mp <- match_parameters(emuDBhandle,
 onTheFlyFunctionName = "forest",onTheFlyParams = list(nominalF1=200),metadata.defaults = list(Age=35,Gender=NA)))
print(mp)

}
