% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_lazy_segment_list.R
\name{needs_collect}
\alias{needs_collect}
\title{Check if Object Needs Materialization}
\usage{
needs_collect(x)
}
\arguments{
\item{x}{An object}
}
\value{
Logical
}
\description{
Check if Object Needs Materialization
}
