% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_methods.R
\name{print.bundle_list}
\alias{print.bundle_list}
\title{Print method for bundle_list with tidyverse-style formatting}
\arguments{
\item{x}{bundle_list object}

\item{...}{Additional arguments (unused)}

\item{n}{Number of rows to show}
}
\description{
Print method for bundle_list with tidyverse-style formatting
}
