% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_list_display.R
\name{print.extended_segment_list}
\alias{print.extended_segment_list}
\alias{.print_extended_segment_list}
\title{Print method for extended_segment_list with pillar formatting}
\usage{
.print_extended_segment_list(x, ..., n = NULL, width = NULL)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments (unused)}

\item{n}{The number of rows to show}

\item{width}{The width of the printed table.}
}
\description{
Print method for extended_segment_list with pillar formatting
}
