% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_lazy_segment_list.R
\name{print.lazy_segment_list}
\alias{print.lazy_segment_list}
\alias{.print_lazy_segment_list}
\title{Print method for lazy_segment_list}
\usage{
.print_lazy_segment_list(x, ..., preview = TRUE)
}
\arguments{
\item{x}{lazy_segment_list object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for lazy_segment_list
}
