% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_lazy_segment_list.R
\name{print.lazy_segment_list}
\alias{print.lazy_segment_list}
\title{Print Method for Lazy Segment List}
\usage{
\method{print}{lazy_segment_list}(x, ...)
}
\arguments{
\item{x}{A lazy_segment_list object}

\item{...}{Additional arguments}
}
\description{
Shows query structure without materializing (unless already cached)
}
