% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_list_display.R
\name{print.segment_list}
\alias{print.segment_list}
\alias{.print_segment_list}
\title{Print method for segment_list with pillar formatting}
\usage{
.print_segment_list(x, ..., n = NULL, width = NULL)
}
\arguments{
\item{x}{The segment_list object}

\item{...}{Additional arguments (unused)}

\item{n}{Number of rows to show}

\item{width}{Width of the printed table}
}
\description{
Print method for segment_list with pillar formatting
}
