% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_signal_extensions_dt.R
\name{process_gender_data}
\alias{process_gender_data}
\title{Process gender-specific data with data.table}
\usage{
process_gender_data(defaults, genders, impute, defaultsEstimatedSampleSize)
}
\description{
Process gender-specific data with data.table
}
\keyword{internal}
