% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_database.R, R/metadata_core.R
\name{process_metadata_list}
\alias{process_metadata_list}
\title{Process and insert metadata list}
\usage{
process_metadata_list(con, db_uuid, session, bundle, meta_list, level)

process_metadata_list(con, db_uuid, session, bundle, meta_list, level)
}
\description{
Process and insert metadata list

Process and insert metadata list into appropriate table - OPTIMIZED BULK INSERT
}
\keyword{internal}
