% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_segment_list.R
\name{quantify}
\alias{quantify}
\title{Quantify generic - Apply DSP to segments}
\usage{
quantify(object, ...)
}
\arguments{
\item{object}{The object to quantify}

\item{...}{Additional arguments passed to methods}
}
\description{
Quantify generic - Apply DSP to segments
}
