% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_database.R, R/metadata_core.R
\name{register_metadata_field}
\alias{register_metadata_field}
\title{Register metadata field}
\usage{
register_metadata_field(con, field_name, field_type)

register_metadata_field(con, field_name, field_type)
}
\description{
Register metadata field

Register a metadata field in the fields tracking table
}
\keyword{internal}
