% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_metadata_optimized.R,
%   R/reindeer-corpus.R
\name{register_metadata_field}
\alias{register_metadata_field}
\title{Register a metadata field in the fields tracking table}
\usage{
register_metadata_field(con, field_name, field_type)

register_metadata_field(con, field_name, field_type)
}
\description{
Register a metadata field in the fields tracking table

Register metadata field
}
\keyword{internal}
