% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer.R
\name{reindeer}
\alias{reindeer}
\title{reindeer: An extension of the capabilities of emuR to work with speech data in
a nordic climate}
\description{
The purpose of this package is to provide a systematic framework for speech
corpora managemnet that retains emuR compatability, but provides provides
expanded ability to store metadata associated with the recording, use the
metadata to guide signal processing when appropriate, and support the user in
future archiving and FAIR use by automating the creation, and continnous
updating, of standard compliant metadata files for corpora.
}
