% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeeR_emuR_re-export.R
\name{rename_bundles}
\alias{rename_bundles}
\title{Rename bundles in emuDB (from emuR)}
\usage{
rename_bundles(emuDBhandle, bundles)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link[emuR]{load_emuDB}}}

\item{bundles}{data.frame like object with the columns
\itemize{
\item \code{session}: name of sessions containing bundle
\item \code{name}: name of bundle
\item \code{name_new}: new name given to bundle
}
It is worth noting that \code{session} and \code{name} are the columns returned by 
\code{\link[emuR]{list_bundles}}.}
}
\description{
Rename bundles of emuDB.
}
\seealso{
\link[emuR:rename_bundles]{emuR::rename_bundles}
}
