% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_database.R, R/metadata_core.R
\name{serialize_metadata_value}
\alias{serialize_metadata_value}
\title{Serialize metadata value}
\usage{
serialize_metadata_value(value)

serialize_metadata_value(value)
}
\description{
Serialize metadata value

Serialize a metadata value for storage
}
\keyword{internal}
