% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR_perspectives.R
\name{set_specOverlay}
\alias{set_specOverlay}
\alias{set_osciOverlay}
\title{Set an SSFF track as an overlay on the spectrogram or oscillograms}
\usage{
set_specOverlay(emuDBhandle, perspective, trackname)

set_osciOverlay(emuDBhandle, perspective, trackname)
}
\arguments{
\item{emuDBhandle}{The database handle.}

\item{perspective}{The name of the perspective in which the overlay should be shown. The perspective should already be defined in the database.}

\item{trackname}{The name of the SSFF track to be set as an overlay on either the oscillogram \link{set_osciOverlay} or spectrogram \link{set_specOverlay}}
}
\description{
The Emu transcription web application allow SSFF tracks to be overlaid on the oscillogram or the spectrogram. Only one track may be set as an overlay
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
reindeer:::create_ae_db() -> ae
#Set up some suitable tracks for overlaying on an oscillogram or spectrogram
remove_ssffTrackDefinition(ae,name="fm",deleteFiles = FALSE)
add_ssffTrackDefinition(ae,name="FORMANTS","fm","fms")
add_ssffTrackDefinition(ae,"rms","rms",onTheFlyFunctionName = "rmsana")

#Set the formants as overlays on the Spectrogram
set_specOverlay(ae,"default","FORMANTS")
#Set the RMS amplitude track as overlays on the waveform (oscillogram)
set_osciOverlay(ae,"default","rms")
serve(ae)
\dontshow{\}) # examplesIf}
}
