% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_methods.R, R/metadata_core.R
\name{[<-.corpus}
\alias{[<-.corpus}
\title{Assign values to corpus bundles - metadata or media import}
\usage{
\method{[}{corpus}(x, i = NULL, j = NULL) <- value
}
\arguments{
\item{x}{A corpus object}

\item{i}{Session name (optional)}

\item{j}{Bundle name (optional)}

\item{value}{Named list of metadata values}

\item{...}{Additional arguments (unused)}
}
\description{
Assign metadata values to a corpus, session, or bundle using bracket notation.
}
\examples{
\dontrun{
mycorpus["Session1", "Bundle1"] <- list(Age = 25, Sex = "Male")
}

}
