% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_segment_list.R
\name{summary.extended_segment_list}
\alias{summary.extended_segment_list}
\title{Summary method for extended_segment_list}
\arguments{
\item{object}{The extended_segment_list object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for extended_segment_list
}
