% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_lazy_segment_list.R
\name{summary.lazy_segment_list}
\alias{summary.lazy_segment_list}
\title{Summary Method for Lazy Segment List}
\usage{
\method{summary}{lazy_segment_list}(object, ...)
}
\arguments{
\item{object}{A lazy_segment_list object}

\item{...}{Additional arguments}
}
\description{
Summary Method for Lazy Segment List
}
