% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindeer_segment_list.R
\name{summary.segment_list}
\alias{summary.segment_list}
\title{Summary method for segment_list}
\arguments{
\item{object}{The segment_list object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for segment_list
}
